refactor: split oversized router, registry, and handler files
