+++ internal/llmclient/client.go
@@ -6,23 +6,13 @@
 package llmclient
 
 import (
-	"bytes"
 	"context"
-	"errors"
-	"fmt"
 	"io"
-	"math"
-	"math/rand"
-	"net"
 	"net/http"
-	"strings"
 	"sync"
 	"time"
 
-	"github.com/goccy/go-json"
-
 	"github.com/enterpilot/gomodel/config"
-	"github.com/enterpilot/gomodel/internal/core"
 	"github.com/enterpilot/gomodel/internal/httpclient"
 )
 
@@ -188,795 +178,4 @@ type Response struct {
 	// provider attempts; it is not relayed to API clients.
 	Header http.Header
 	Body   []byte
-}
-
-// maxErrorBodyBytes caps how much of an upstream error body is read into
-// memory. It matches core's audit capture cap; a misbehaving upstream that
-// answers an error status with an endless body must not be buffered whole.
-const maxErrorBodyBytes = 64 * 1024
-
-// attachResponseHeaders records the upstream response headers on a provider
-// GatewayError so failed attempts can be audited. It is a no-op for other
-// error types or a nil header set.
-func attachResponseHeaders(err error, header http.Header) error {
-	if header == nil {
-		return err
-	}
-	var gatewayErr *core.GatewayError
-	if errors.As(err, &gatewayErr) && gatewayErr != nil {
-		gatewayErr.ResponseHeaders = header.Clone()
-	}
-	return err
-}
-
-type requestScope struct {
-	ctx           context.Context
-	startedAt     time.Time
-	requestInfo   RequestInfo
-	halfOpenProbe bool
-}
-
-func (c *Client) beginRequest(ctx context.Context, req Request, stream bool) (requestScope, error) {
-	scope := requestScope{
-		ctx:       ctx,
-		startedAt: time.Now(),
-		requestInfo: RequestInfo{
-			Provider:        c.config.ProviderName,
-			Model:           requestModel(req),
-			Operation:       req.Operation,
-			Endpoint:        req.Endpoint,
-			Method:          req.Method,
-			Stream:          stream,
-			StreamUncertain: req.StreamUncertain,
-		},
-	}
-
-	if c.config.Hooks.OnRequestStart != nil {
-		scope.ctx = c.config.Hooks.OnRequestStart(scope.ctx, scope.requestInfo)
-	}
-
-	if c.circuitBreaker != nil {
-		allowed, probe := c.circuitBreaker.acquire()
-		if !allowed {
-			err := core.NewProviderError(c.config.ProviderName, http.StatusServiceUnavailable,
-				"circuit breaker is open - provider temporarily unavailable", nil)
-			c.finishRequest(scope, http.StatusServiceUnavailable, err)
-			return requestScope{}, err
-		}
-		scope.halfOpenProbe = probe
-	}
-
-	return scope, nil
-}
-
-func requestModel(req Request) string {
-	if model := strings.TrimSpace(req.Model); model != "" {
-		return model
-	}
-	return extractModel(req.Body)
-}
-
-func (c *Client) finishRequest(scope requestScope, statusCode int, err error) {
-	if c.config.Hooks.OnRequestEnd == nil {
-		return
-	}
-	circuitState := ""
-	if c.circuitBreaker != nil {
-		circuitState = c.circuitBreaker.State()
-	}
-	c.config.Hooks.OnRequestEnd(scope.ctx, ResponseInfo{
-		Provider:        c.config.ProviderName,
-		ProviderType:    scope.requestInfo.ProviderType,
-		Model:           scope.requestInfo.Model,
-		Operation:       scope.requestInfo.Operation,
-		Endpoint:        scope.requestInfo.Endpoint,
-		Method:          scope.requestInfo.Method,
-		StatusCode:      statusCode,
-		Duration:        time.Since(scope.startedAt),
-		Stream:          scope.requestInfo.Stream,
-		StreamUncertain: scope.requestInfo.StreamUncertain,
-		Error:           err,
-		CircuitState:    circuitState,
-	})
-}
-
-func (c *Client) finishStreamFirstChunk(scope requestScope, statusCode int) {
-	if c.config.Hooks.OnStreamFirstChunk == nil {
-		return
-	}
-	c.config.Hooks.OnStreamFirstChunk(scope.ctx, ResponseInfo{
-		Provider:        c.config.ProviderName,
-		ProviderType:    scope.requestInfo.ProviderType,
-		Model:           scope.requestInfo.Model,
-		Operation:       scope.requestInfo.Operation,
-		Endpoint:        scope.requestInfo.Endpoint,
-		Method:          scope.requestInfo.Method,
-		StatusCode:      statusCode,
-		Duration:        time.Since(scope.startedAt),
-		Stream:          true,
-		StreamUncertain: scope.requestInfo.StreamUncertain,
-	})
-}
-
-func (c *Client) finishStreamEmpty(scope requestScope, statusCode int, err error) {
-	if c.config.Hooks.OnStreamEmpty == nil {
-		return
-	}
-	c.config.Hooks.OnStreamEmpty(scope.ctx, ResponseInfo{
-		Provider:     c.config.ProviderName,
-		ProviderType: scope.requestInfo.ProviderType,
-		Model:        scope.requestInfo.Model,
-		Operation:    scope.requestInfo.Operation,
-		Endpoint:     scope.requestInfo.Endpoint,
-		Method:       scope.requestInfo.Method,
-		StatusCode:   statusCode,
-		Duration:     time.Since(scope.startedAt),
-		Stream:       true,
-		Error:        err,
-	})
-}
-
-func (c *Client) observeFirstChunk(scope requestScope, resp *http.Response, stream bool) {
-	if resp == nil || resp.Body == nil || !stream {
-		return
-	}
-	resp.Body = &firstChunkReadCloser{
-		ReadCloser: resp.Body,
-		onFirstChunk: func() {
-			c.finishStreamFirstChunk(scope, resp.StatusCode)
-		},
-		onEmpty: func(err error) {
-			c.finishStreamEmpty(scope, resp.StatusCode, err)
-		},
-	}
-}
-
-// firstChunkReadCloser reports the first moment a stream body delivers bytes,
-// or that it ended before delivering any. Exactly one of the two fires.
-type firstChunkReadCloser struct {
-	io.ReadCloser
-	once         sync.Once
-	onFirstChunk func()
-	onEmpty      func(err error)
-}
-
-func (r *firstChunkReadCloser) Read(p []byte) (int, error) {
-	n, err := r.ReadCloser.Read(p)
-	switch {
-	case n > 0:
-		r.once.Do(r.onFirstChunk)
-	case err != nil:
-		r.once.Do(func() { r.onEmpty(err) })
-	}
-	return n, err
-}
-
-// completeScope is the standard terminal step for a request that has passed
-// beginRequest. It records the circuit-breaker outcome (using cbErr to decide
-// whether the failure was transport-level) and emits the metrics observation.
-// Use this whenever a code path returns from one of the public Do* methods.
-func (c *Client) completeScope(scope requestScope, statusCode int, err, cbErr error) {
-	c.recordCircuitBreakerCompletion(scope, statusCode, cbErr)
-	c.finishRequest(scope, statusCode, err)
-}
-
-// finishRequestWithoutBreaker finalises a request that never reached the
-// upstream (local request-build errors): no breaker outcome is recorded, but
-// a consumed half-open probe slot must still be returned or the breaker would
-// reject all traffic forever.
-func (c *Client) finishRequestWithoutBreaker(scope requestScope, statusCode int, err error) {
-	c.releaseHalfOpenProbe(scope)
-	c.finishRequest(scope, statusCode, err)
-}
-
-// releaseHalfOpenProbe frees the breaker's probe slot when this request held
-// it but ended without a success/failure verdict.
-func (c *Client) releaseHalfOpenProbe(scope requestScope) {
-	if c.circuitBreaker != nil && scope.halfOpenProbe {
-		c.circuitBreaker.releaseProbe()
-	}
-}
-
-// failAfterRetries handles the "exhausted retries with no captured error"
-// fallback shared by the retrying entry points (DoRaw, DoPassthrough). The
-// returned error is also reported through the scope.
-func (c *Client) failAfterRetries(scope requestScope) error {
-	err := core.NewProviderError(c.config.ProviderName, http.StatusBadGateway, "request failed after retries", nil)
-	c.completeScope(scope, http.StatusBadGateway, err, err)
-	return err
-}
-
-// waitForRetryAttempt sleeps for the per-attempt backoff (a no-op for
-// attempt 0) and finalises the scope if the context cancels mid-wait. The
-// caller should return early when this returns a non-nil error.
-func (c *Client) waitForRetryAttempt(ctx context.Context, scope requestScope, attempt int) error {
-	if err := c.waitForRetry(ctx, attempt); err != nil {
-		c.finishRequest(scope, 0, err)
-		return err
-	}
-	return nil
-}
-
-func (c *Client) recordCircuitBreakerCompletion(scope requestScope, statusCode int, err error) {
-	if c.circuitBreaker == nil {
-		return
-	}
-	if err != nil {
-		// A caller-side cancellation aborts the transport but proves nothing
-		// about provider health, so it is neither a success nor a failure.
-		// Client deadlines (context.DeadlineExceeded) still count: the
-		// provider failed to answer within the latency budget.
-		if errors.Is(err, context.Canceled) {
-			c.releaseHalfOpenProbe(scope)
-			return
-		}
-		c.circuitBreaker.RecordFailure()
-		return
-	}
-	if statusCode == http.StatusTooManyRequests {
-		if c.circuitBreaker.IsHalfOpen() {
-			c.circuitBreaker.RecordFailure()
-		}
-		return
-	}
-	if c.shouldTripCircuitBreaker(statusCode) {
-		c.circuitBreaker.RecordFailure()
-		return
-	}
-	c.circuitBreaker.RecordSuccess()
-}
-
-func (c *Client) shouldTripCircuitBreaker(statusCode int) bool {
-	if statusCode == http.StatusTooManyRequests {
-		return false
-	}
-	return c.isRetryable(statusCode) || statusCode >= http.StatusInternalServerError
-}
-
-func (c *Client) maxAttempts() int {
-	maxAttempts := c.config.Retry.MaxRetries + 1
-	if maxAttempts < 1 {
-		return 1
-	}
-	return maxAttempts
-}
-
-func (c *Client) waitForRetry(ctx context.Context, attempt int) error {
-	if attempt <= 0 {
-		return nil
-	}
-	backoff := c.calculateBackoff(attempt)
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-time.After(backoff):
-		return nil
-	}
-}
-
-// Do executes a request with retries and circuit breaking, then unmarshals the response
-func (c *Client) Do(ctx context.Context, req Request, result any) error {
-	resp, err := c.DoRaw(ctx, req)
-	if err != nil {
-		return err
-	}
-
-	if result != nil {
-		if err := json.Unmarshal(resp.Body, result); err != nil {
-			return core.NewProviderError(c.config.ProviderName, http.StatusBadGateway, "failed to unmarshal response: "+err.Error(), err)
-		}
-	}
-
-	return nil
-}
-
-// DoRaw executes a request with retries and circuit breaking, returning the raw response.
-//
-// # Metrics Behavior
-//
-// Metrics hooks (OnRequestStart/OnRequestEnd) are called at this level to track logical
-// requests from the caller's perspective, not individual retry attempts. This ensures:
-//
-//   - Request counts reflect user-facing requests, not internal HTTP calls
-//   - Duration metrics include total time across all retries (useful for SLOs)
-//   - In-flight gauge accurately reflects concurrent logical requests
-//
-// Behavior comparison (hooks at DoRaw vs per-attempt):
-//
-//	| Scenario                             | Per-attempt (old)           | DoRaw level (current)            |
-//	|--------------------------------------|-----------------------------|----------------------------------|
-//	| 1 request, succeeds first try        | 1 observation               | 1 observation                    |
-//	| 1 request, fails twice then succeeds | 3 observations              | 1 observation (success)          |
-//	| 1 request, fails all 3 retries       | 3 observations              | 1 observation (error)            |
-//	| Duration metric                      | Each attempt's duration     | Total duration including retries |
-//	| In-flight gauge                      | Bounces up/down per attempt | Accurate concurrent count        |
-//
-// The final status code and error in metrics reflect the outcome after all retry attempts.
-func (c *Client) DoRaw(ctx context.Context, req Request) (*Response, error) {
-	scope, err := c.beginRequest(ctx, req, false)
-	if err != nil {
-		closeRawBodyReader(req)
-		return nil, err
-	}
-	ctx = scope.ctx
-
-	var lastErr error
-	var lastStatusCode int
-	lastErrFromTransport := false
-	maxAttempts := c.maxAttempts()
-	if req.RawBodyReader != nil {
-		maxAttempts = 1
-	}
-
-	for attempt := 0; attempt < maxAttempts; attempt++ {
-		if err := c.waitForRetryAttempt(ctx, scope, attempt); err != nil {
-			closeRawBodyReader(req)
-			return nil, err
-		}
-
-		resp, err := c.doRequest(ctx, req)
-		if err != nil {
-			lastErr = err
-			lastStatusCode = extractStatusCode(err)
-			// Caller-side build errors (validation, body conflicts, marshal
-			// failures) will repeat deterministically and never reached the
-			// upstream — short-circuit without retrying and skip the breaker
-			// entirely (a 400 with cbErr=nil would otherwise be recorded as
-			// a success by recordCircuitBreakerCompletion).
-			if isLocalRequestBuildError(err) {
-				c.finishRequestWithoutBreaker(scope, lastStatusCode, err)
-				return nil, err
-			}
-			lastErrFromTransport = true
-			// Client-side timeouts are already the caller's latency budget. Do
-			// not retry them, or the logical request can outlive HTTP_TIMEOUT.
-			if scope.halfOpenProbe || isClientTimeoutGatewayError(lastErr) {
-				c.completeScope(scope, lastStatusCode, lastErr, lastErr)
-				return nil, lastErr
-			}
-			continue
-		}
-
-		// Check for retryable status codes
-		if c.isRetryable(resp.StatusCode) {
-			lastErr = attachResponseHeaders(core.ParseProviderError(c.config.ProviderName, resp.StatusCode, resp.Body, nil), resp.Header)
-			lastStatusCode = resp.StatusCode
-			lastErrFromTransport = false
-			if scope.halfOpenProbe {
-				c.completeScope(scope, lastStatusCode, lastErr, nil)
-				return nil, lastErr
-			}
-			continue
-		}
-
-		// Non-retryable error
-		if resp.StatusCode != http.StatusOK {
-			parsedErr := attachResponseHeaders(core.ParseProviderError(c.config.ProviderName, resp.StatusCode, resp.Body, nil), resp.Header)
-			c.completeScope(scope, resp.StatusCode, parsedErr, nil)
-			return nil, parsedErr
-		}
-
-		// Success
-		c.completeScope(scope, resp.StatusCode, nil, nil)
-		return resp, nil
-	}
-
-	// All retries exhausted
-	if lastErr != nil {
-		var circuitErr error
-		if lastErrFromTransport {
-			circuitErr = lastErr
-		}
-		c.completeScope(scope, lastStatusCode, lastErr, circuitErr)
-		return nil, lastErr
-	}
-	return nil, c.failAfterRetries(scope)
-}
-
-// DoStream executes a streaming request, returning a ReadCloser
-// Note: Streaming requests do NOT retry (as partial data may have been sent)
-// Metrics note: Duration is measured from start to stream establishment, not stream close
-func (c *Client) DoStream(ctx context.Context, req Request) (io.ReadCloser, error) {
-	scope, err := c.beginRequest(ctx, req, true)
-	if err != nil {
-		closeRawBodyReader(req)
-		return nil, err
-	}
-
-	resp, err := c.doHTTPRequest(scope.ctx, req)
-	if err != nil {
-		statusCode := extractStatusCode(err)
-		// Caller-side build errors never reached the upstream — skip the
-		// breaker entirely so neither RecordFailure nor RecordSuccess fires.
-		if isLocalRequestBuildError(err) {
-			c.finishRequestWithoutBreaker(scope, statusCode, err)
-			return nil, err
-		}
-		c.completeScope(scope, statusCode, err, err)
-		return nil, err
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		respBody, readErr := io.ReadAll(io.LimitReader(resp.Body, maxErrorBodyBytes))
-		if readErr != nil {
-			respBody = []byte("failed to read error response")
-		}
-		_ = resp.Body.Close()
-
-		providerErr := attachResponseHeaders(core.ParseProviderError(c.config.ProviderName, resp.StatusCode, respBody, nil), resp.Header)
-		c.completeScope(scope, resp.StatusCode, providerErr, nil)
-		return nil, providerErr
-	}
-
-	// The stream can outlive the request by minutes while transport internals
-	// keep resp.Request reachable. GetBody closes over the fully marshaled
-	// request payload; redirects and transparent transport retries only
-	// consult it inside Do, so dropping it here releases the payload for the
-	// stream's lifetime without changing behavior.
-	if resp.Request != nil {
-		resp.Request.GetBody = nil
-	}
-
-	c.completeScope(scope, resp.StatusCode, nil, nil)
-	c.observeFirstChunk(scope, resp, true)
-	return resp.Body, nil
-}
-
-func canRetryPassthrough(req Request) bool {
-	if req.RawBodyReader != nil {
-		return false
-	}
-	if hasIdempotencyKey(req.Headers) {
-		return true
-	}
-	switch strings.ToUpper(strings.TrimSpace(req.Method)) {
-	case http.MethodGet, http.MethodHead, http.MethodOptions, http.MethodPut:
-		return true
-	default:
-		return false
-	}
-}
-
-func hasIdempotencyKey(headers http.Header) bool {
-	for key, values := range headers {
-		if http.CanonicalHeaderKey(strings.TrimSpace(key)) != "Idempotency-Key" {
-			continue
-		}
-		for _, value := range values {
-			if strings.TrimSpace(value) != "" {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// DoPassthrough executes a request and returns the raw upstream HTTP response.
-// Unlike DoRaw, it preserves non-200 responses for the caller to proxy unchanged.
-func (c *Client) DoPassthrough(ctx context.Context, req Request) (*http.Response, error) {
-	stream := req.Stream || strings.Contains(strings.ToLower(strings.Join(req.Headers.Values("Accept"), ",")), "text/event-stream")
-	scope, err := c.beginRequest(ctx, req, stream)
-	if err != nil {
-		closeRawBodyReader(req)
-		return nil, err
-	}
-	ctx = scope.ctx
-
-	maxAttempts := 1
-	if canRetryPassthrough(req) {
-		maxAttempts = c.maxAttempts()
-	}
-
-	for attempt := 0; attempt < maxAttempts; attempt++ {
-		if err := c.waitForRetryAttempt(ctx, scope, attempt); err != nil {
-			closeRawBodyReader(req)
-			return nil, err
-		}
-
-		resp, err := c.doHTTPRequest(ctx, req)
-		if err != nil {
-			statusCode := extractStatusCode(err)
-			// Caller-side build errors will repeat and never hit the upstream;
-			// skip the breaker entirely (cbErr=nil would otherwise record a
-			// spurious success for a 400-class status).
-			if isLocalRequestBuildError(err) {
-				c.finishRequestWithoutBreaker(scope, statusCode, err)
-				return nil, err
-			}
-			if scope.halfOpenProbe || isClientTimeoutGatewayError(err) || attempt == maxAttempts-1 {
-				c.completeScope(scope, statusCode, err, err)
-				return nil, err
-			}
-			continue
-		}
-
-		retryable := c.isRetryable(resp.StatusCode)
-		if retryable {
-			if scope.halfOpenProbe || attempt == maxAttempts-1 {
-				c.completeScope(scope, resp.StatusCode, nil, nil)
-				return resp, nil
-			}
-			_ = resp.Body.Close()
-			continue
-		}
-
-		if resp.StatusCode >= http.StatusOK && resp.StatusCode < http.StatusMultipleChoices {
-			// The response settles what a bounded body peek could not: an
-			// uncertain request that came back as JSON completes here as a
-			// buffered call, so hooks see the resolved stream state.
-			responseStream := stream || strings.Contains(strings.ToLower(resp.Header.Get("Content-Type")), "text/event-stream")
-			scope.requestInfo.Stream = responseStream
-			scope.requestInfo.StreamUncertain = false
-			c.completeScope(scope, resp.StatusCode, nil, nil)
-			c.observeFirstChunk(scope, resp, responseStream)
-			return resp, nil
-		}
-		c.completeScope(scope, resp.StatusCode, nil, nil)
-		return resp, nil
-	}
-
-	return nil, c.failAfterRetries(scope)
-}
-
-// extractModel attempts to extract the model name from a request body
-// UnknownModel is the model label reported for requests whose model cannot be
-// recovered from the request body (body-less discovery GETs, availability
-// probes, multipart uploads). Hook consumers that attribute traffic per model
-// should treat it as "not model-attributed".
-const UnknownModel = "unknown"
-
-func extractModel(body any) string {
-	if body == nil {
-		return UnknownModel
-	}
-
-	// Try ChatRequest
-	if chatReq, ok := body.(*core.ChatRequest); ok && chatReq != nil {
-		return chatReq.Model
-	}
-
-	// Try ResponsesRequest
-	if respReq, ok := body.(*core.ResponsesRequest); ok && respReq != nil {
-		return respReq.Model
-	}
-
-	// Try AudioSpeechRequest. Transcription has no JSON body (multipart upload),
-	// so its model cannot be recovered here and stays "unknown".
-	if speechReq, ok := body.(*core.AudioSpeechRequest); ok && speechReq != nil {
-		return speechReq.Model
-	}
-
-	// Unknown request type
-	return UnknownModel
-}
-
-// extractStatusCode tries to extract HTTP status code from an error
-func extractStatusCode(err error) int {
-	if gwErr, ok := errors.AsType[*core.GatewayError](err); ok {
-		return gwErr.StatusCode
-	}
-
-	// Network or unknown error
-	return 0
-}
-
-// doHTTPRequest executes a single HTTP request without retries and returns the
-// live upstream response. Metrics hooks are called at the logical request
-// level, not here, to avoid counting each attempt separately.
-func (c *Client) doHTTPRequest(ctx context.Context, req Request) (*http.Response, error) {
-	httpReq, err := c.buildRequest(ctx, req)
-	if err != nil {
-		// The transport owns closing the body once the request reaches it;
-		// a request that never gets there must release the reader here.
-		closeRawBodyReader(req)
-		return nil, err
-	}
-
-	resp, err := c.httpClient.Do(httpReq)
-	if err != nil {
-		return nil, core.NewProviderError(c.config.ProviderName, providerErrorStatusCode(err), "failed to send request: "+err.Error(), err)
-	}
-	return resp, nil
-}
-
-// closeRawBodyReader releases a caller-supplied streaming body when the
-// request fails before reaching the HTTP transport, which otherwise closes it
-// on every path. Pipe-backed uploads (files, audio transcription) rely on
-// this to unblock their producer goroutines instead of leaking them — and the
-// upload buffers they pin — for the process lifetime.
-func closeRawBodyReader(req Request) {
-	if closer, ok := req.RawBodyReader.(io.Closer); ok {
-		_ = closer.Close()
-	}
-}
-
-// doRequest executes a single HTTP request without retries.
-// Note: Metrics hooks are called at the DoRaw level, not here, to avoid
-// counting each retry attempt as a separate request.
-func (c *Client) doRequest(ctx context.Context, req Request) (*Response, error) {
-	resp, err := c.doHTTPRequest(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		_ = resp.Body.Close()
-	}()
-
-	// Successful responses are read whole (large results are legitimate);
-	// error bodies are bounded — they only feed error parsing and audit
-	// capture, both of which cap at the same size.
-	reader := io.Reader(resp.Body)
-	if resp.StatusCode >= http.StatusBadRequest {
-		reader = io.LimitReader(resp.Body, maxErrorBodyBytes)
-	}
-	body, err := io.ReadAll(reader)
-	if err != nil {
-		return nil, core.NewProviderError(c.config.ProviderName, providerErrorStatusCode(err), "failed to read response: "+err.Error(), err)
-	}
-
-	return &Response{
-		StatusCode:  resp.StatusCode,
-		ContentType: resp.Header.Get("Content-Type"),
-		Header:      resp.Header,
-		Body:        body,
-	}, nil
-}
-
-// buildRequest creates an HTTP request from a Request
-func (c *Client) buildRequest(ctx context.Context, req Request) (*http.Request, error) {
-	// Validate request
-	if req.Method == "" {
-		return nil, core.NewInvalidRequestError("HTTP method is required", nil)
-	}
-	if req.Endpoint == "" {
-		return nil, core.NewInvalidRequestError("endpoint is required", nil)
-	}
-
-	// Validate HTTP method
-	switch req.Method {
-	case http.MethodGet, http.MethodPost, http.MethodPut, http.MethodPatch, http.MethodDelete, http.MethodHead, http.MethodOptions:
-		// Valid methods
-	default:
-		return nil, core.NewInvalidRequestError(fmt.Sprintf("invalid HTTP method: %s", req.Method), nil)
-	}
-
-	url := c.BaseURL() + req.Endpoint
-
-	var bodyReader io.Reader
-	bodySources := 0
-	if req.Body != nil {
-		bodySources++
-	}
-	if req.RawBody != nil {
-		bodySources++
-	}
-	if req.RawBodyReader != nil {
-		bodySources++
-	}
-	if bodySources > 1 {
-		return nil, core.NewInvalidRequestError("Body, RawBody, and RawBodyReader are mutually exclusive", nil)
-	}
-	if req.RawBodyReader != nil {
-		bodyReader = req.RawBodyReader
-	} else if req.RawBody != nil {
-		bodyReader = bytes.NewReader(req.RawBody)
-	} else if req.Body != nil {
-		bodyBytes, err := json.Marshal(req.Body)
-		if err != nil {
-			return nil, core.NewInvalidRequestError("failed to marshal request", err)
-		}
-		bodyReader = bytes.NewReader(bodyBytes)
-	}
-
-	httpReq, err := http.NewRequestWithContext(ctx, req.Method, url, bodyReader)
-	if err != nil {
-		return nil, core.NewInvalidRequestError("failed to create request", err)
-	}
-
-	// Set default content type for requests with body
-	if req.Body != nil {
-		httpReq.Header.Set("Content-Type", "application/json")
-	}
-
-	// Apply provider-specific headers
-	if c.headerSetter != nil {
-		c.headerSetter(httpReq)
-	}
-
-	// Apply request-specific headers
-	for key, values := range req.Headers {
-		httpReq.Header.Del(key)
-		for _, value := range values {
-			httpReq.Header.Add(key, value)
-		}
-	}
-
-	return httpReq, nil
-}
-
-// calculateBackoff calculates the backoff duration for a given attempt with jitter
-func (c *Client) calculateBackoff(attempt int) time.Duration {
-	retry := c.config.Retry
-	backoff := float64(retry.InitialBackoff) * math.Pow(retry.BackoffFactor, float64(attempt-1))
-	if backoff > float64(retry.MaxBackoff) {
-		backoff = float64(retry.MaxBackoff)
-	}
-
-	if retry.JitterFactor > 0 {
-		jitter := backoff * retry.JitterFactor
-		//nolint:gosec // math/rand is fine for jitter, no crypto needed
-		backoff = backoff - jitter + (rand.Float64() * 2 * jitter)
-	}
-
-	return time.Duration(backoff)
-}
-
-// isRetryable returns true if the status code indicates a retryable error
-func (c *Client) isRetryable(statusCode int) bool {
-	// Retry on rate limits and specific server errors that are typically transient
-	return statusCode == http.StatusTooManyRequests ||
-		statusCode == http.StatusServiceUnavailable ||
-		statusCode == http.StatusBadGateway ||
-		statusCode == http.StatusGatewayTimeout
-}
-
-func providerErrorStatusCode(err error) int {
-	if isTimeoutError(err) {
-		return http.StatusGatewayTimeout
-	}
-	return http.StatusBadGateway
-}
-
-func isTimeoutError(err error) bool {
-	if err == nil {
-		return false
-	}
-	if errors.Is(err, context.DeadlineExceeded) {
-		return true
-	}
-
-	var netErr net.Error
-	if errors.As(err, &netErr) && netErr.Timeout() {
-		return true
-	}
-
-	message := strings.ToLower(err.Error())
-	return strings.Contains(message, "client.timeout exceeded") ||
-		strings.Contains(message, "timeout awaiting response headers")
-}
-
-// isLocalRequestBuildError reports whether err originated from buildRequest
-// (e.g. an empty endpoint, an invalid HTTP method, mutually-exclusive bodies,
-// or a marshal failure). Such errors are caller-side: they will repeat
-// deterministically on retry and must not be charged to the circuit breaker
-// because the upstream provider was never contacted.
-//
-// buildRequest is the only producer of *core.GatewayError with type
-// ErrorTypeInvalidRequest along the doRequest/doHTTPRequest path — the other
-// transport-layer wrappers all use NewProviderError. ParseProviderError runs
-// only on a returned response body, so an InvalidRequest seen in the
-// transport-error branch can only have come from buildRequest.
-func isLocalRequestBuildError(err error) bool {
-	if err == nil {
-		return false
-	}
-	var gatewayErr *core.GatewayError
-	if !errors.As(err, &gatewayErr) || gatewayErr == nil {
-		return false
-	}
-	return gatewayErr.Type == core.ErrorTypeInvalidRequest
-}
-
-func isClientTimeoutGatewayError(err error) bool {
-	var gatewayErr *core.GatewayError
-	if !errors.As(err, &gatewayErr) || gatewayErr == nil {
-		return isTimeoutError(err)
-	}
-	if gatewayErr.StatusCode != http.StatusGatewayTimeout {
-		return false
-	}
-	if isTimeoutError(gatewayErr.Err) {
-		return true
-	}
-	return isTimeoutError(gatewayErr)
 }
