+++ internal/providers/gemini/gemini.go
@@ -3,20 +3,13 @@ package gemini
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"fmt"
-	"io"
-	"log/slog"
 	"net/http"
-	"net/url"
 	"os"
-	"slices"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/goccy/go-json"
 	"golang.org/x/sync/singleflight"
 
 	"github.com/enterpilot/gomodel/internal/core"
@@ -87,12 +80,6 @@ type Provider struct {
 	cacheMu      sync.Mutex
 	cacheObjects map[string]geminiCacheObject
 	cacheFlight  singleflight.Group
-}
-
-type geminiCacheObject struct {
-	name       string
-	expiresAt  time.Time
-	retryAfter time.Time
 }
 
 // New creates a new Gemini provider.
@@ -409,640 +396,4 @@ func nativeBaseURLFromOpenAICompatibleBaseURL(baseURL string) (string, bool) {
 		return "", false
 	}
 	return nativeBaseURL, true
-}
-
-// adaptChatRequest rewrites a ChatRequest for Gemini's OpenAI-compatible endpoint.
-// Gemini uses "reasoning_effort" as a top-level string (e.g. "low", "medium", "high"),
-// not the nested "reasoning": {"effort": "..."} format.
-func adaptChatRequest(req *core.ChatRequest) (*core.ChatRequest, error) {
-	if req.Reasoning == nil || req.Reasoning.Effort == "" {
-		return req, nil
-	}
-	return providers.AdaptReasoningEffortRequest(req, req.Reasoning.Effort)
-}
-
-func (p *Provider) openAICompatibleChatBody(req *core.ChatRequest) (any, error) {
-	if p.backend == geminiBackendVertex {
-		if model := vertexOpenAIModelID(req.Model); strings.TrimSpace(model) != "" {
-			rewritten := *req
-			rewritten.Model = model
-			req = &rewritten
-		}
-	}
-	return adaptChatRequest(req)
-}
-
-func (p *Provider) openAICompatibleEmbeddingBody(req *core.EmbeddingRequest) (any, error) {
-	if p.backend != geminiBackendVertex {
-		return req, nil
-	}
-	model := vertexOpenAIModelID(req.Model)
-	if strings.TrimSpace(model) == "" {
-		return req, nil
-	}
-	rewritten := *req
-	rewritten.Model = model
-	return &rewritten, nil
-}
-
-// ChatCompletion sends a chat completion request to Gemini
-func (p *Provider) ChatCompletion(ctx context.Context, req *core.ChatRequest) (*core.ChatResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	if req == nil {
-		return nil, core.NewInvalidRequestError("chat request is required", nil)
-	}
-	if p.useNativeAPI {
-		return p.nativeChatCompletion(ctx, req)
-	}
-	body, err := p.openAICompatibleChatBody(req)
-	if err != nil {
-		return nil, err
-	}
-	var resp core.ChatResponse
-	err = p.client.Do(ctx, llmclient.Request{
-		Method:    http.MethodPost,
-		Endpoint:  "/chat/completions",
-		Operation: llmclient.OperationChat,
-		Model:     req.Model,
-		Body:      body,
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	core.EnsureModel(&resp.Model, req.Model)
-	if resp.Provider == "" {
-		resp.Provider = p.responseProviderName()
-	}
-	return &resp, nil
-}
-
-func (p *Provider) nativeChatCompletion(ctx context.Context, req *core.ChatRequest) (*core.ChatResponse, error) {
-	body, err := convertChatRequestToGemini(req)
-	if err != nil {
-		return nil, err
-	}
-	p.prepareCachedContent(ctx, req, body)
-	var geminiResp geminiGenerateContentResponse
-	err = p.nativeClient.Do(ctx, llmclient.Request{
-		Method:    http.MethodPost,
-		Endpoint:  nativeGenerateEndpoint(req.Model),
-		Operation: llmclient.OperationGenerateContent,
-		Model:     req.Model,
-		Body:      body,
-	}, &geminiResp)
-	if err != nil {
-		return nil, err
-	}
-	return nativeChatResponse(req, &geminiResp, p.responseProviderName())
-}
-
-// StreamChatCompletion returns a raw response body for streaming (caller must close)
-func (p *Provider) StreamChatCompletion(ctx context.Context, req *core.ChatRequest) (io.ReadCloser, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	if req == nil {
-		return nil, core.NewInvalidRequestError("chat request is required", nil)
-	}
-	if p.useNativeAPI {
-		return p.nativeStreamChatCompletion(ctx, req)
-	}
-	streamReq := req.WithStreaming()
-	body, err := p.openAICompatibleChatBody(streamReq)
-	if err != nil {
-		return nil, err
-	}
-	stream, err := p.client.DoStream(ctx, llmclient.Request{
-		Method:    http.MethodPost,
-		Endpoint:  "/chat/completions",
-		Operation: llmclient.OperationChat,
-		Model:     req.Model,
-		Body:      body,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// Gemini's OpenAI-compatible endpoint returns OpenAI-format SSE, so we can pass it through directly
-	return stream, nil
-}
-
-func (p *Provider) nativeStreamChatCompletion(ctx context.Context, req *core.ChatRequest) (io.ReadCloser, error) {
-	streamReq := req.WithStreaming()
-	body, err := convertChatRequestToGemini(streamReq)
-	if err != nil {
-		return nil, err
-	}
-	p.prepareCachedContent(ctx, req, body)
-	stream, err := p.nativeClient.DoStream(ctx, llmclient.Request{
-		Method:    http.MethodPost,
-		Endpoint:  nativeStreamEndpoint(req.Model),
-		Operation: llmclient.OperationGenerateContent,
-		Model:     req.Model,
-		Body:      body,
-	})
-	if err != nil {
-		return nil, err
-	}
-	includeUsage := streamReq.StreamOptions != nil && streamReq.StreamOptions.IncludeUsage
-	return newGeminiNativeStream(stream, req.Model, includeUsage, p.responseProviderName()), nil
-}
-
-type geminiCreateCachedContentRequest struct {
-	Model             string          `json:"model"`
-	SystemInstruction *geminiContent  `json:"systemInstruction,omitempty"`
-	Contents          []geminiContent `json:"contents"`
-	Tools             []geminiTool    `json:"tools,omitempty"`
-	TTL               string          `json:"ttl"`
-}
-
-type geminiCreateCachedContentResponse struct {
-	Name       string    `json:"name"`
-	ExpireTime time.Time `json:"expireTime"`
-}
-
-// prepareCachedContent materializes the stable prefix selected by the
-// post-routing planner. Cache creation is best-effort: an unsupported model or
-// endpoint must never turn a request that would otherwise work into a failure.
-func (p *Provider) prepareCachedContent(ctx context.Context, req *core.ChatRequest, body *geminiGenerateContentRequest) {
-	if body == nil || body.CachedContent != "" || p.backend != geminiBackendAIStudio || len(body.Contents) == 0 {
-		return
-	}
-	if req.PromptCachePlan == nil || strings.TrimSpace(req.PromptCachePlan.Key) == "" {
-		return
-	}
-	// The final content is the live turn. A system instruction, tools, or an
-	// earlier content item must remain before it for a useful cached prefix.
-	if len(body.Contents) == 1 && body.SystemInstruction == nil && len(body.Tools) == 0 {
-		return
-	}
-	key, ok := p.scopedCachedContentKey(ctx, req.PromptCachePlan.Key)
-	if !ok {
-		return
-	}
-	if cached, suppress := p.cachedContentObject(key, time.Now()); suppress {
-		if cached != "" {
-			useGeminiCachedPrefix(body, cached)
-		}
-		return
-	}
-
-	value, _, _ := p.cacheFlight.Do(key, func() (any, error) {
-		now := time.Now()
-		if cached, suppress := p.cachedContentObject(key, now); suppress {
-			return cached, nil
-		}
-		// Cache creation may outlive the request that happened to lead the
-		// singleflight. Keep affinity values while detaching cancellation.
-		createCtx, cancel := context.WithTimeout(context.WithoutCancel(ctx), geminiCacheCreateTimeout)
-		defer cancel()
-		createReq := geminiCreateCachedContentRequest{
-			Model:             "models/" + normalizeGeminiModelID(req.Model),
-			SystemInstruction: body.SystemInstruction,
-			Contents:          append([]geminiContent(nil), body.Contents[:len(body.Contents)-1]...),
-			Tools:             append([]geminiTool(nil), body.Tools...),
-			TTL:               fmt.Sprintf("%ds", geminiCacheTTL/time.Second),
-		}
-		var created geminiCreateCachedContentResponse
-		err := p.nativeClient.Do(createCtx, llmclient.Request{
-			Method: http.MethodPost, Endpoint: "/cachedContents", Body: &createReq,
-		}, &created)
-		now = time.Now()
-		if err != nil || strings.TrimSpace(created.Name) == "" {
-			if err == nil {
-				err = fmt.Errorf("cached-content creation returned an empty name")
-			}
-			slog.DebugContext(ctx, "Gemini cached-content creation skipped", "model", req.Model, "error", err)
-			p.storeCachedContentObject(key, geminiCacheObject{retryAfter: now.Add(geminiCacheFailureBackoff)}, now)
-			return "", nil
-		}
-		expiresAt := created.ExpireTime
-		if expiresAt.IsZero() {
-			expiresAt = now.Add(geminiCacheTTL)
-		}
-		if !now.Add(geminiCacheFreshness).Before(expiresAt) {
-			p.storeCachedContentObject(key, geminiCacheObject{retryAfter: now.Add(geminiCacheFailureBackoff)}, now)
-			return "", nil
-		}
-		p.storeCachedContentObject(key, geminiCacheObject{name: created.Name, expiresAt: expiresAt}, now)
-		return created.Name, nil
-	})
-	if cached, ok := value.(string); ok && cached != "" {
-		useGeminiCachedPrefix(body, cached)
-	}
-}
-
-func (p *Provider) scopedCachedContentKey(ctx context.Context, planKey string) (string, bool) {
-	credential, stable := p.keys.StableForContext(ctx)
-	if !stable {
-		return "", false
-	}
-	digest := sha256.Sum256([]byte(credential))
-	return planKey + ":" + hex.EncodeToString(digest[:]), true
-}
-
-func (p *Provider) cachedContentObject(key string, now time.Time) (string, bool) {
-	p.cacheMu.Lock()
-	defer p.cacheMu.Unlock()
-	entry, ok := p.cacheObjects[key]
-	if !ok {
-		return "", false
-	}
-	if entry.name == "" {
-		if now.Before(entry.retryAfter) {
-			return "", true
-		}
-		delete(p.cacheObjects, key)
-		return "", false
-	}
-	if now.Add(geminiCacheFreshness).After(entry.expiresAt) {
-		delete(p.cacheObjects, key)
-		return "", false
-	}
-	return entry.name, true
-}
-
-func (p *Provider) storeCachedContentObject(key string, entry geminiCacheObject, now time.Time) {
-	p.cacheMu.Lock()
-	defer p.cacheMu.Unlock()
-	if p.cacheObjects == nil {
-		p.cacheObjects = make(map[string]geminiCacheObject)
-	}
-	for candidate, cached := range p.cacheObjects {
-		if (cached.name == "" && !now.Before(cached.retryAfter)) ||
-			(cached.name != "" && now.Add(geminiCacheFreshness).After(cached.expiresAt)) {
-			delete(p.cacheObjects, candidate)
-		}
-	}
-	if len(p.cacheObjects) >= geminiCacheObjectLimit {
-		for candidate := range p.cacheObjects {
-			delete(p.cacheObjects, candidate)
-			break
-		}
-	}
-	p.cacheObjects[key] = entry
-}
-
-func useGeminiCachedPrefix(body *geminiGenerateContentRequest, name string) {
-	body.CachedContent = name
-	body.Contents = append([]geminiContent(nil), body.Contents[len(body.Contents)-1])
-	body.SystemInstruction = nil
-	body.Tools = nil
-}
-
-// geminiModel represents a model in Gemini's native API response
-type geminiModel struct {
-	Name             string   `json:"name"`
-	DisplayName      string   `json:"displayName"`
-	Description      string   `json:"description"`
-	SupportedMethods []string `json:"supportedGenerationMethods"`
-	InputTokenLimit  int      `json:"inputTokenLimit"`
-	OutputTokenLimit int      `json:"outputTokenLimit"`
-	Temperature      *float64 `json:"temperature,omitempty"`
-	TopP             *float64 `json:"topP,omitempty"`
-	TopK             *int     `json:"topK,omitempty"`
-}
-
-// geminiModelsResponse represents the native Gemini models list response
-type geminiModelsResponse struct {
-	Models          []geminiModel `json:"models"`
-	PublisherModels []geminiModel `json:"publisherModels"`
-}
-
-func geminiModelSupportedMethods(modelID string, methods []string) (supportsGenerate, supportsEmbed, supportsImage bool) {
-	normalized := normalizeGeminiModelID(modelID)
-	if len(methods) == 0 {
-		isEmbedding := strings.HasPrefix(normalized, "text-embedding-") || strings.HasPrefix(normalized, "gemini-embedding-")
-		return strings.HasPrefix(normalized, "gemini-") && !isEmbedding, isEmbedding,
-			strings.HasPrefix(normalized, "imagen-")
-	}
-	supportsGenerate = slices.Contains(methods, "generateContent") || slices.Contains(methods, "streamGenerateContent")
-	// Imagen models list only the predict method; Gemini image models carry
-	// generateContent, so their image capability is inferred from the ID.
-	supportsImage = (strings.HasPrefix(normalized, "imagen-") && slices.Contains(methods, "predict")) ||
-		(supportsGenerate && isGeminiImageModelID(normalized))
-	return supportsGenerate, slices.Contains(methods, "embedContent"), supportsImage
-}
-
-// isGeminiImageModelID recognizes generateContent models with image output
-// (gemini-2.5-flash-image, gemini-2.0-flash-preview-image-generation,
-// gemini-3-pro-image-preview, ...) by the -image marker Google uses in their
-// IDs. This only seeds discovery metadata; registry enrichment overrides it.
-func isGeminiImageModelID(normalized string) bool {
-	return strings.HasPrefix(normalized, "gemini-") &&
-		(strings.Contains(normalized, "-image-") || strings.HasSuffix(normalized, "-image"))
-}
-
-// geminiDiscoveredMetadata stamps modes/categories from the native listing's
-// supportedGenerationMethods so embedding models are classified even when the
-// remote model registry has no entry (new or preview IDs). Registry enrichment
-// replaces this metadata whenever it does have an entry, and operator config
-// merges on top, so the discovery stamp is only the lowest-precedence signal.
-func geminiDiscoveredMetadata(supportsGenerate, supportsEmbed, supportsImage bool) *core.ModelMetadata {
-	modes := make([]string, 0, 3)
-	if supportsGenerate {
-		modes = append(modes, "chat")
-	}
-	if supportsEmbed {
-		modes = append(modes, "embedding")
-	}
-	if supportsImage {
-		modes = append(modes, "image_generation")
-		if supportsGenerate {
-			// Only generateContent image models accept input images to edit;
-			// Imagen predict models generate from text alone.
-			modes = append(modes, "image_edit")
-		}
-	}
-	if len(modes) == 0 {
-		return nil
-	}
-	return &core.ModelMetadata{
-		Modes:      modes,
-		Categories: core.CategoriesForModes(modes),
-	}
-}
-
-// ListModels retrieves the list of available models from Gemini
-func (p *Provider) ListModels(ctx context.Context) (*core.ModelsResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	modelsClient := p.modelsClient
-	if modelsClient == nil {
-		modelsClient = p.nativeClient
-	}
-	rawResp, err := modelsClient.DoRaw(ctx, llmclient.Request{
-		Method:   http.MethodGet,
-		Endpoint: "/models",
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	now := time.Now().Unix()
-
-	// Preferred path: native Gemini models response.
-	// If the payload contains an explicit "models" field with an empty array,
-	// return an empty list instead of falling through to fallback parsing.
-	var nativeProbe struct {
-		Models          json.RawMessage `json:"models"`
-		PublisherModels json.RawMessage `json:"publisherModels"`
-	}
-	if err := json.Unmarshal(rawResp.Body, &nativeProbe); err == nil && (nativeProbe.Models != nil || nativeProbe.PublisherModels != nil) {
-		var geminiResp geminiModelsResponse
-		if err := json.Unmarshal(rawResp.Body, &geminiResp); err != nil {
-			return nil, core.NewProviderError(p.responseProviderName(), http.StatusBadGateway, "failed to parse native Gemini models response", err)
-		}
-		modelEntries := append(geminiResp.Models, geminiResp.PublisherModels...)
-		if len(modelEntries) == 0 {
-			return &core.ModelsResponse{
-				Object: "list",
-				Data:   []core.Model{},
-			}, nil
-		}
-
-		models := make([]core.Model, 0, len(modelEntries))
-
-		for _, gm := range modelEntries {
-			modelID := displayModelIDFromGemini(gm.Name, p.backend)
-
-			supportsGenerate, supportsEmbed, supportsImage := geminiModelSupportedMethods(modelID, gm.SupportedMethods)
-
-			if (supportsGenerate || supportsEmbed || supportsImage) && isGeminiExposedModel(modelID) {
-				models = append(models, core.Model{
-					ID:       modelID,
-					Object:   "model",
-					OwnedBy:  "google",
-					Created:  now,
-					Metadata: geminiDiscoveredMetadata(supportsGenerate, supportsEmbed, supportsImage),
-				})
-			}
-		}
-
-		return &core.ModelsResponse{
-			Object: "list",
-			Data:   models,
-		}, nil
-	}
-
-	// Fallback path: OpenAI-compatible models list.
-	var openAIResp core.ModelsResponse
-	if err := json.Unmarshal(rawResp.Body, &openAIResp); err == nil && openAIResp.Object == "list" {
-		models := make([]core.Model, 0, len(openAIResp.Data))
-		for _, m := range openAIResp.Data {
-			modelID := displayModelIDFromGemini(m.ID, p.backend)
-			isOpenAICompatModel := isGeminiExposedModel(modelID)
-			if !isOpenAICompatModel {
-				continue
-			}
-			models = append(models, core.Model{
-				ID:      modelID,
-				Object:  "model",
-				OwnedBy: "google",
-				Created: now,
-			})
-		}
-		return &core.ModelsResponse{
-			Object: "list",
-			Data:   models,
-		}, nil
-	}
-
-	responsePreview := string(rawResp.Body)
-	if len(responsePreview) > 512 {
-		responsePreview = responsePreview[:512] + "...(truncated)"
-	}
-	return nil, core.NewProviderError(p.responseProviderName(), http.StatusBadGateway, "unexpected Gemini models response format", fmt.Errorf("models response body: %s", responsePreview))
-}
-
-// Responses sends a Responses API request to Gemini (converted to chat format)
-func (p *Provider) Responses(ctx context.Context, req *core.ResponsesRequest) (*core.ResponsesResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	return providers.ResponsesViaChat(ctx, p, req)
-}
-
-// Embeddings sends an embeddings request to Gemini: the native
-// batchEmbedContents API in native mode, or the OpenAI-compatible endpoint
-// otherwise. The Vertex backend always uses the OpenAI-compatible surface —
-// Vertex has no batchEmbedContents; its native prediction path is served by
-// the dedicated vertex provider.
-func (p *Provider) Embeddings(ctx context.Context, req *core.EmbeddingRequest) (*core.EmbeddingResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	if req == nil {
-		return nil, core.NewInvalidRequestError("embedding request is required", nil)
-	}
-	if p.useNativeAPI && p.backend == geminiBackendAIStudio {
-		return p.nativeEmbeddings(ctx, req)
-	}
-	body, err := p.openAICompatibleEmbeddingBody(req)
-	if err != nil {
-		return nil, err
-	}
-	var resp core.EmbeddingResponse
-	err = p.client.Do(ctx, llmclient.Request{
-		Method:    http.MethodPost,
-		Endpoint:  "/embeddings",
-		Operation: llmclient.OperationEmbeddings,
-		Model:     req.Model,
-		Body:      body,
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	core.EnsureModel(&resp.Model, req.Model)
-	if resp.Provider == "" {
-		resp.Provider = p.responseProviderName()
-	}
-	return &resp, nil
-}
-
-// StreamResponses returns a raw response body for streaming Responses API (caller must close)
-func (p *Provider) StreamResponses(ctx context.Context, req *core.ResponsesRequest) (io.ReadCloser, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	return providers.StreamResponsesViaChat(ctx, p, req, p.responseProviderName())
-}
-
-// CreateBatch creates a native Gemini batch job through its OpenAI-compatible endpoint.
-func (p *Provider) CreateBatch(ctx context.Context, req *core.BatchRequest) (*core.BatchResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	var resp core.BatchResponse
-	err := p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/batches",
-		Body:     req,
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	providers.EnsureProviderBatchID(&resp)
-	return &resp, nil
-}
-
-// GetBatch retrieves a native Gemini batch job.
-func (p *Provider) GetBatch(ctx context.Context, id string) (*core.BatchResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	var resp core.BatchResponse
-	err := p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodGet,
-		Endpoint: "/batches/" + url.PathEscape(id),
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	providers.EnsureProviderBatchID(&resp)
-	return &resp, nil
-}
-
-// ListBatches lists native Gemini batch jobs.
-func (p *Provider) ListBatches(ctx context.Context, limit int, after string) (*core.BatchListResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	endpoint := providers.PaginatedEndpoint("/batches", limit, "after", after)
-
-	var resp core.BatchListResponse
-	err := p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodGet,
-		Endpoint: endpoint,
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	providers.EnsureProviderBatchIDs(&resp)
-	return &resp, nil
-}
-
-// CancelBatch cancels a native Gemini batch job.
-func (p *Provider) CancelBatch(ctx context.Context, id string) (*core.BatchResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	var resp core.BatchResponse
-	err := p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/batches/" + url.PathEscape(id) + "/cancel",
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	providers.EnsureProviderBatchID(&resp)
-	return &resp, nil
-}
-
-// GetBatchResults fetches Gemini batch results via the output file API.
-func (p *Provider) GetBatchResults(ctx context.Context, id string) (*core.BatchResultsResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	return providers.FetchBatchResultsFromOutputFile(ctx, p.client, p.responseProviderName(), id)
-}
-
-// CreateFile uploads a file through Gemini's OpenAI-compatible /files API.
-func (p *Provider) CreateFile(ctx context.Context, req *core.FileCreateRequest) (*core.FileObject, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	resp, err := providers.CreateOpenAICompatibleFile(ctx, p.client, req)
-	if err != nil {
-		return nil, err
-	}
-	resp.Provider = p.responseProviderName()
-	return resp, nil
-}
-
-// ListFiles lists files through Gemini's OpenAI-compatible /files API.
-func (p *Provider) ListFiles(ctx context.Context, purpose string, limit int, after string) (*core.FileListResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	resp, err := providers.ListOpenAICompatibleFiles(ctx, p.client, purpose, limit, after)
-	if err != nil {
-		return nil, err
-	}
-	for i := range resp.Data {
-		resp.Data[i].Provider = p.responseProviderName()
-	}
-	return resp, nil
-}
-
-// GetFile retrieves one file object through Gemini's OpenAI-compatible /files API.
-func (p *Provider) GetFile(ctx context.Context, id string) (*core.FileObject, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	resp, err := providers.GetOpenAICompatibleFile(ctx, p.client, id)
-	if err != nil {
-		return nil, err
-	}
-	resp.Provider = p.responseProviderName()
-	return resp, nil
-}
-
-// DeleteFile deletes a file object through Gemini's OpenAI-compatible /files API.
-func (p *Provider) DeleteFile(ctx context.Context, id string) (*core.FileDeleteResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	return providers.DeleteOpenAICompatibleFile(ctx, p.client, id)
-}
-
-// GetFileContent fetches raw file bytes through Gemini's /files/{id}/content API.
-func (p *Provider) GetFileContent(ctx context.Context, id string) (*core.FileContentResponse, error) {
-	if err := p.ready(); err != nil {
-		return nil, err
-	}
-	return providers.GetOpenAICompatibleFileContent(ctx, p.client, id)
 }
