+++ internal/providers/registry.go
@@ -4,11 +4,8 @@ package providers
 import (
 	"fmt"
 	"maps"
-	"slices"
-	"sort"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/goccy/go-json"
 
@@ -156,86 +153,6 @@ func (r *ModelRegistry) invalidateSortedCaches() {
 	r.categoryCache = nil
 	r.qualifiedByName = nil
 	r.qualifiedByType = nil
-}
-
-// ResolveProviderSelector resolves a qualified "<segment>/<modelID>" selector,
-// where segment is a provider instance name or a provider type, to the concrete
-// provider-name-qualified selector. Provider-name matches take precedence over
-// provider-type matches, mirroring catalog-scan resolution. Returns ok=false
-// when the segment+model pair is not a direct name/type match so callers can
-// fall back to slower resolution for raw slash-shaped IDs and other edge cases.
-//
-// This is O(1) and exists so the per-request routing path does not copy and
-// linearly scan the entire model catalog.
-func (r *ModelRegistry) ResolveProviderSelector(segment, modelID string) (core.ModelSelector, bool) {
-	segment = strings.TrimSpace(segment)
-	modelID = strings.TrimSpace(modelID)
-	if segment == "" || modelID == "" {
-		return core.ModelSelector{}, false
-	}
-	key := segment + "/" + modelID
-
-	r.mu.RLock()
-	if r.qualifiedByName != nil {
-		sel, ok := lookupSelectorIndex(r.qualifiedByName, r.qualifiedByType, key)
-		r.mu.RUnlock()
-		return sel, ok
-	}
-	r.mu.RUnlock()
-
-	r.mu.Lock()
-	r.buildSelectorIndexLocked()
-	sel, ok := lookupSelectorIndex(r.qualifiedByName, r.qualifiedByType, key)
-	r.mu.Unlock()
-	return sel, ok
-}
-
-func lookupSelectorIndex(byName, byType map[string]core.ModelSelector, key string) (core.ModelSelector, bool) {
-	if sel, ok := byName[key]; ok {
-		return sel, true
-	}
-	if sel, ok := byType[key]; ok {
-		return sel, true
-	}
-	return core.ModelSelector{}, false
-}
-
-// buildSelectorIndexLocked populates the qualified selector index from the
-// current catalog. Caller must hold the write lock. On provider-type collisions
-// it keeps the lexicographically smallest provider name so resolution is
-// deterministic and matches the previous sorted-scan behavior.
-func (r *ModelRegistry) buildSelectorIndexLocked() {
-	if r.qualifiedByName != nil {
-		return
-	}
-	total := 0
-	for _, providerModels := range r.modelsByProvider {
-		total += len(providerModels)
-	}
-	byName := make(map[string]core.ModelSelector, total)
-	byType := make(map[string]core.ModelSelector, total)
-	for providerName, providerModels := range r.modelsByProvider {
-		for _, info := range providerModels {
-			publicName := providerName
-			if info.ProviderName != "" {
-				publicName = info.ProviderName
-			}
-			id := info.Model.ID
-			if publicName == "" || id == "" {
-				continue
-			}
-			sel := core.ModelSelector{Provider: publicName, Model: id}
-			byName[publicName+"/"+id] = sel
-			if info.ProviderType != "" {
-				typeKey := info.ProviderType + "/" + id
-				if existing, ok := byType[typeKey]; !ok || sel.Provider < existing.Provider {
-					byType[typeKey] = sel
-				}
-			}
-		}
-	}
-	r.qualifiedByName = byName
-	r.qualifiedByType = byType
 }
 
 // RegisterProvider adds a provider to the registry
@@ -398,775 +315,4 @@ func (r *ModelRegistry) UnregisterProvider(providerName string) {
 	delete(r.discoveredByProvider, providerName)
 	r.publishFilteredInventoryLocked()
 	r.invalidateSortedCaches()
-}
-
-// lookupLocked resolves a model selector to its registry entry. Qualified
-// selectors ("<provider>/<model>") are looked up under the configured provider
-// name first; a qualified selector naming a configured provider that does not
-// serve the model is a definitive miss. Otherwise the slash may be part of the
-// model ID itself (e.g. "meta-llama/Meta-Llama-3-70B"), so the raw selector is
-// tried against the global map. Callers must hold r.mu.
-func (r *ModelRegistry) lookupLocked(model string) (*ModelInfo, bool) {
-	model = strings.TrimSpace(model)
-	providerName, modelID := splitModelSelector(model)
-	if providerName != "" {
-		if providerModels, ok := r.modelsByProvider[providerName]; ok {
-			if info, exists := providerModelInfo(providerModels, modelID, model); exists {
-				return info, true
-			}
-		}
-		if r.hasConfiguredProviderNameLocked(providerName) {
-			return nil, false
-		}
-	}
-
-	info, ok := r.models[model]
-	return info, ok
-}
-
-// GetProvider returns the provider for the given model, or nil if not found
-func (r *ModelRegistry) GetProvider(model string) core.Provider {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	if info, ok := r.lookupLocked(model); ok {
-		return info.Provider
-	}
-	return nil
-}
-
-// GetModel returns the registry-backed model info for the given model, or nil if not found.
-// Callers must treat the returned data as read-only.
-func (r *ModelRegistry) GetModel(model string) *ModelInfo {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	if info, ok := r.lookupLocked(model); ok {
-		return info
-	}
-	return nil
-}
-
-// LookupModel returns a shallow copy of the concrete model for the given selector.
-// Qualified selectors use the configured provider name prefix when present.
-func (r *ModelRegistry) LookupModel(model string) (*core.Model, bool) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	if info, ok := r.lookupLocked(model); ok {
-		cloned := info.Model
-		return &cloned, true
-	}
-	return nil, false
-}
-
-// Supports returns true if the registry has a provider for the given model
-func (r *ModelRegistry) Supports(model string) bool {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	_, ok := r.lookupLocked(model)
-	return ok
-}
-
-// ModelAvailable reports whether the model is registered AND its provider's
-// inventory is fresh (latest refresh succeeded). Virtual-model load balancing
-// uses this to skip providers whose upstream is failing, while Supports keeps
-// resolving stale models so direct requests still reach the provider and fail
-// with an honest 502/503 instead of "model not found".
-func (r *ModelRegistry) ModelAvailable(model string) bool {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	info, ok := r.lookupLocked(model)
-	if !ok {
-		return false
-	}
-	return !r.providerRuntime[info.ProviderName].inventoryStale
-}
-
-// providerAdvertisedLocked reports whether providerName's models belong in
-// model listings. A provider whose latest refresh or availability probe
-// failed (inventoryStale) keeps its carried-forward inventory resolvable for
-// direct requests, but offline providers must not advertise models in
-// GET /v1/models, the dashboard, or failover candidate selection.
-// Caller must hold r.mu.
-func (r *ModelRegistry) providerAdvertisedLocked(providerName string) bool {
-	return !r.providerRuntime[providerName].inventoryStale
-}
-
-// ListModels returns all advertised models in the registry, sorted by model ID
-// for consistent ordering. Models whose owning provider's inventory is stale
-// are excluded (see providerAdvertisedLocked).
-// The sorted slice is cached and rebuilt only when the underlying models change.
-// Returns a defensive copy so callers cannot mutate the internal cache.
-func (r *ModelRegistry) ListModels() []core.Model {
-	r.mu.RLock()
-	if cached := r.sortedModels; cached != nil {
-		r.mu.RUnlock()
-		return append([]core.Model(nil), cached...)
-	}
-	r.mu.RUnlock()
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	// Double-check: another goroutine may have built it while we waited for the lock.
-	if r.sortedModels != nil {
-		return append([]core.Model(nil), r.sortedModels...)
-	}
-
-	models := make([]core.Model, 0, len(r.models))
-	for _, info := range r.models {
-		if !r.providerAdvertisedLocked(info.ProviderName) {
-			continue
-		}
-		models = append(models, info.Model)
-	}
-	sort.Slice(models, func(i, j int) bool { return models[i].ID < models[j].ID })
-
-	r.sortedModels = models
-	return append([]core.Model(nil), models...)
-}
-
-// ListPublicModels returns all provider-backed models as public selectors in
-// providerName/modelID form, sorted by public model ID. Models the owning
-// provider cannot actually serve (audio-only models on providers without audio
-// support) and models from providers with stale inventory are not advertised.
-func (r *ModelRegistry) ListPublicModels() []core.Model {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	total := 0
-	for _, models := range r.modelsByProvider {
-		total += len(models)
-	}
-
-	result := make([]core.Model, 0, total)
-	for providerName, models := range r.modelsByProvider {
-		if !r.providerAdvertisedLocked(providerName) {
-			continue
-		}
-		for modelID, info := range models {
-			if !providerCanServeModel(info) {
-				continue
-			}
-			model := info.Model
-			model.ID = qualifyPublicModelID(providerName, modelID)
-			model.OwnedBy = providerName
-			result = append(result, model)
-		}
-	}
-	sort.Slice(result, func(i, j int) bool { return result[i].ID < result[j].ID })
-	return result
-}
-
-// ListUnqualifiedPublicModels returns advertised models under their bare
-// model IDs, sorted by ID. When several providers expose the same model ID,
-// only the provider an unqualified request resolves to (the global catalog
-// entry, see rebuildGlobalModelMap) is listed, so the response never
-// advertises an ID that would route somewhere else. OwnedBy carries the
-// provider name so callers can still tell which provider serves the model.
-func (r *ModelRegistry) ListUnqualifiedPublicModels() []core.Model {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	result := make([]core.Model, 0, len(r.models))
-	for modelID, info := range r.models {
-		if !r.providerAdvertisedLocked(info.ProviderName) || !providerCanServeModel(info) {
-			continue
-		}
-		model := info.Model
-		model.ID = modelID
-		model.OwnedBy = info.ProviderName
-		result = append(result, model)
-	}
-	sort.Slice(result, func(i, j int) bool { return result[i].ID < result[j].ID })
-	return result
-}
-
-// providerCanServeModel reports whether the owning provider implements the
-// capabilities a model needs. Upstream inventories and the metadata registry
-// can list audio-only (TTS/STT) or image-only models for providers whose
-// gateway adapter has no audio or image support; advertising those invites
-// calls that can only fail with "does not support audio operations" or "does
-// not support image generation". Models without mode metadata are kept —
-// missing data must not hide a model.
-func providerCanServeModel(info *ModelInfo) bool {
-	if isAudioOnlyModel(info.Model) {
-		_, ok := info.Provider.(core.AudioProvider)
-		return ok
-	}
-	if isImageOnlyModel(info.Model) {
-		_, ok := info.Provider.(core.ImageProvider)
-		return ok
-	}
-	return true
-}
-
-// isAudioOnlyModel reports whether every declared mode is an audio operation.
-func isAudioOnlyModel(model core.Model) bool {
-	return onlyHasModes(model, "audio_speech", "audio_transcription")
-}
-
-// isImageOnlyModel reports whether every declared mode is an image operation.
-func isImageOnlyModel(model core.Model) bool {
-	return onlyHasModes(model, "image_generation", "image_edit")
-}
-
-func onlyHasModes(model core.Model, allowed ...string) bool {
-	if model.Metadata == nil || len(model.Metadata.Modes) == 0 {
-		return false
-	}
-	for _, mode := range model.Metadata.Modes {
-		if !slices.Contains(allowed, mode) {
-			return false
-		}
-	}
-	return true
-}
-
-// ModelCount returns the number of registered models
-func (r *ModelRegistry) ModelCount() int {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return len(r.models)
-}
-
-// GetProviderType returns the provider type string for the given model.
-// Returns empty string if the model is not found.
-func (r *ModelRegistry) GetProviderType(model string) string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	if info, ok := r.lookupLocked(model); ok {
-		return info.ProviderType
-	}
-	return ""
-}
-
-// GetProviderName returns the concrete configured provider instance name for
-// the given model selector. Returns empty string if the model is not found.
-func (r *ModelRegistry) GetProviderName(model string) string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	if info, ok := r.lookupLocked(model); ok {
-		return info.ProviderName
-	}
-	return ""
-}
-
-// GetProviderNameForType returns the first registered configured provider name
-// for the given provider type. This follows the same first-registered behavior
-// used when provider-typed routes resolve a concrete provider instance.
-func (r *ModelRegistry) GetProviderNameForType(providerType string) string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	providerType = strings.TrimSpace(providerType)
-	if providerType == "" {
-		return ""
-	}
-	for _, provider := range r.providers {
-		if r.providerTypes[provider] != providerType {
-			continue
-		}
-		return r.providerNames[provider]
-	}
-	return ""
-}
-
-// GetProviderTypeForName returns the provider type for the given concrete
-// configured provider instance name.
-func (r *ModelRegistry) GetProviderTypeForName(providerName string) string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	providerName = strings.TrimSpace(providerName)
-	if providerName == "" {
-		return ""
-	}
-	for _, provider := range r.providers {
-		if r.providerNames[provider] != providerName {
-			continue
-		}
-		return r.providerTypes[provider]
-	}
-	return ""
-}
-
-// ProviderByType returns the first registered provider for the given provider type.
-// This lookup is independent of discovered models so provider-typed routes keep
-// working even when a provider currently exposes zero models.
-func (r *ModelRegistry) ProviderByType(providerType string) core.Provider {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	providerType = strings.TrimSpace(providerType)
-	if providerType == "" {
-		return nil
-	}
-	for _, provider := range r.providers {
-		if r.providerTypes[provider] == providerType {
-			return provider
-		}
-	}
-	return nil
-}
-
-// ProviderByName returns the registered provider for a configured provider
-// instance name.
-func (r *ModelRegistry) ProviderByName(providerName string) core.Provider {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	providerName = strings.TrimSpace(providerName)
-	if providerName == "" {
-		return nil
-	}
-	for _, provider := range r.providers {
-		if r.providerNames[provider] == providerName {
-			return provider
-		}
-	}
-	return nil
-}
-
-// ProviderTypes returns the unique registered provider types in sorted order.
-// This inventory is independent of discovered models.
-func (r *ModelRegistry) ProviderTypes() []string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	seen := make(map[string]struct{}, len(r.providerTypes))
-	result := make([]string, 0, len(r.providerTypes))
-	for _, provider := range r.providers {
-		providerType := r.providerTypes[provider]
-		if providerType == "" {
-			continue
-		}
-		if _, exists := seen[providerType]; exists {
-			continue
-		}
-		seen[providerType] = struct{}{}
-		result = append(result, providerType)
-	}
-	sort.Strings(result)
-	return result
-}
-
-// ProviderNames returns the configured provider instance names in registration order.
-func (r *ModelRegistry) ProviderNames() []string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	result := make([]string, 0, len(r.providers))
-	for _, provider := range r.providers {
-		providerName := r.providerNames[provider]
-		if providerName == "" {
-			continue
-		}
-		result = append(result, providerName)
-	}
-	return result
-}
-
-// splitModelSelector splits an already-trimmed selector into its provider and
-// model segments. Whitespace around the slash is tolerated.
-func splitModelSelector(model string) (providerName, modelID string) {
-	if model == "" {
-		return "", ""
-	}
-	before, after, found := strings.Cut(model, "/")
-	if !found {
-		return "", model
-	}
-	providerName = strings.TrimSpace(before)
-	modelID = strings.TrimSpace(after)
-	if providerName == "" || modelID == "" {
-		return "", model
-	}
-	return providerName, modelID
-}
-
-func providerModelInfo(providerModels map[string]*ModelInfo, modelID, rawModel string) (*ModelInfo, bool) {
-	if info, exists := providerModels[modelID]; exists {
-		return info, true
-	}
-	if rawModel != "" && rawModel != modelID {
-		if info, exists := providerModels[rawModel]; exists {
-			return info, true
-		}
-	}
-	return nil, false
-}
-
-func qualifyPublicModelID(providerName, modelID string) string {
-	if providerName == "" {
-		return modelID
-	}
-	if modelID == "" {
-		return providerName
-	}
-	return providerName + "/" + modelID
-}
-
-func (r *ModelRegistry) hasConfiguredProviderNameLocked(providerName string) bool {
-	if providerName == "" {
-		return false
-	}
-	for _, configuredName := range r.providerNames {
-		if configuredName == providerName {
-			return true
-		}
-	}
-	return false
-}
-
-// ModelWithProvider holds a model alongside provider metadata and its public selector.
-type ModelWithProvider struct {
-	Model        core.Model `json:"model"`
-	ProviderType string     `json:"provider_type"`
-	ProviderName string     `json:"provider_name"`
-	Selector     string     `json:"selector"`
-}
-
-// ListModelsWithProvider returns all provider-backed models with provider metadata,
-// sorted by public selector.
-// The sorted slice is cached and rebuilt only when the underlying models change.
-// Returns a defensive copy so callers cannot mutate the internal cache.
-func (r *ModelRegistry) ListModelsWithProvider() []ModelWithProvider {
-	r.mu.RLock()
-	if cached := r.sortedModelsWithProvider; cached != nil {
-		r.mu.RUnlock()
-		return append([]ModelWithProvider(nil), cached...)
-	}
-	r.mu.RUnlock()
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if r.sortedModelsWithProvider != nil {
-		return append([]ModelWithProvider(nil), r.sortedModelsWithProvider...)
-	}
-
-	total := 0
-	for _, providerModels := range r.modelsByProvider {
-		total += len(providerModels)
-	}
-
-	result := make([]ModelWithProvider, 0, total)
-	for providerName, providerModels := range r.modelsByProvider {
-		if !r.providerAdvertisedLocked(providerName) {
-			continue
-		}
-		for modelID, info := range providerModels {
-			publicProviderName := providerName
-			if info.ProviderName != "" {
-				publicProviderName = info.ProviderName
-			}
-			result = append(result, ModelWithProvider{
-				Model:        info.Model,
-				ProviderType: info.ProviderType,
-				ProviderName: publicProviderName,
-				Selector:     qualifyPublicModelID(publicProviderName, modelID),
-			})
-		}
-	}
-	sort.Slice(result, func(i, j int) bool { return result[i].Selector < result[j].Selector })
-
-	r.sortedModelsWithProvider = result
-	return append([]ModelWithProvider(nil), result...)
-}
-
-// cacheableCategory reports whether category is a known value that should be cached.
-// CategoryAll is handled separately (delegates to ListModelsWithProvider).
-var cacheableCategories = map[core.ModelCategory]struct{}{
-	core.CategoryTextGeneration: {},
-	core.CategoryEmbedding:      {},
-	core.CategoryImage:          {},
-	core.CategoryAudio:          {},
-	core.CategoryVideo:          {},
-	core.CategoryUtility:        {},
-}
-
-// ListModelsWithProviderByCategory returns provider-backed models filtered by
-// category, sorted by public selector.
-// If category is CategoryAll, returns all models (same as ListModelsWithProvider).
-// Results for known categories are cached and rebuilt only when the underlying models change.
-// Returns a defensive copy so callers cannot mutate the internal cache.
-func (r *ModelRegistry) ListModelsWithProviderByCategory(category core.ModelCategory) []ModelWithProvider {
-	if category == core.CategoryAll {
-		return r.ListModelsWithProvider()
-	}
-
-	_, cacheable := cacheableCategories[category]
-
-	if cacheable {
-		r.mu.RLock()
-		if r.categoryCache != nil {
-			if cached, ok := r.categoryCache[category]; ok {
-				r.mu.RUnlock()
-				return append([]ModelWithProvider(nil), cached...)
-			}
-		}
-		r.mu.RUnlock()
-	}
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if cacheable && r.categoryCache != nil {
-		if cached, ok := r.categoryCache[category]; ok {
-			return append([]ModelWithProvider(nil), cached...)
-		}
-	}
-
-	result := make([]ModelWithProvider, 0)
-	for providerName, providerModels := range r.modelsByProvider {
-		if !r.providerAdvertisedLocked(providerName) {
-			continue
-		}
-		for modelID, info := range providerModels {
-			if info.Model.Metadata == nil || !hasCategory(info.Model.Metadata.Categories, category) {
-				continue
-			}
-			result = append(result, ModelWithProvider{
-				Model:        info.Model,
-				ProviderType: info.ProviderType,
-				ProviderName: info.ProviderName,
-				Selector:     qualifyPublicModelID(info.ProviderName, modelID),
-			})
-		}
-	}
-	sort.Slice(result, func(i, j int) bool { return result[i].Selector < result[j].Selector })
-
-	if cacheable {
-		if r.categoryCache == nil {
-			r.categoryCache = make(map[core.ModelCategory][]ModelWithProvider)
-		}
-		r.categoryCache[category] = result
-	}
-	return result
-}
-
-// hasCategory returns true if the category slice contains the target category.
-func hasCategory(cats []core.ModelCategory, target core.ModelCategory) bool {
-	return slices.Contains(cats, target)
-}
-
-// CategoryCount holds a model category and the number of models in it.
-type CategoryCount struct {
-	Category    core.ModelCategory `json:"category"`
-	DisplayName string             `json:"display_name"`
-	Count       int                `json:"count"`
-}
-
-// categoryDisplayNames maps categories to human-readable display names.
-var categoryDisplayNames = map[core.ModelCategory]string{
-	core.CategoryAll:            "All",
-	core.CategoryTextGeneration: "Text Generation",
-	core.CategoryEmbedding:      "Embeddings",
-	core.CategoryImage:          "Image",
-	core.CategoryAudio:          "Audio",
-	core.CategoryVideo:          "Video",
-	core.CategoryUtility:        "Utility",
-}
-
-// GetCategoryCounts returns model counts per category, in display order.
-// A model with multiple categories is counted in each.
-func (r *ModelRegistry) GetCategoryCounts() []CategoryCount {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	counts := make(map[core.ModelCategory]int)
-	total := 0
-	for providerName, providerModels := range r.modelsByProvider {
-		if !r.providerAdvertisedLocked(providerName) {
-			continue
-		}
-		for _, info := range providerModels {
-			total++
-			if info.Model.Metadata != nil {
-				for _, cat := range info.Model.Metadata.Categories {
-					counts[cat]++
-				}
-			}
-		}
-	}
-
-	allCategories := core.AllCategories()
-	result := make([]CategoryCount, 0, len(allCategories))
-	for _, cat := range allCategories {
-		count := counts[cat]
-		if cat == core.CategoryAll {
-			count = total
-		}
-		displayName := categoryDisplayNames[cat]
-		if displayName == "" {
-			displayName = string(cat)
-		}
-		result = append(result, CategoryCount{
-			Category:    cat,
-			DisplayName: displayName,
-			Count:       count,
-		})
-	}
-	return result
-}
-
-// ProviderCount returns the number of registered providers
-func (r *ModelRegistry) ProviderCount() int {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return len(r.providers)
-}
-
-// failureMessage returns the trimmed text of a failure for runtime state. An
-// empty stored message means "no failure", so a message that is empty or only
-// whitespace becomes a fixed marker instead of reading as success.
-func failureMessage(message string) string {
-	if trimmed := strings.TrimSpace(message); trimmed != "" {
-		return trimmed
-	}
-	return "unknown error"
-}
-
-// optionalFailureMessage is failureMessage for fields where "" means no
-// failure occurred.
-func optionalFailureMessage(message string) string {
-	if message == "" {
-		return ""
-	}
-	return failureMessage(message)
-}
-
-// RecordAvailabilityCheck stores the latest startup or explicit availability
-// probe result for a configured provider name.
-func (r *ModelRegistry) RecordAvailabilityCheck(providerName string, err error) {
-	providerName = strings.TrimSpace(providerName)
-	if providerName == "" {
-		return
-	}
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	state := r.providerRuntime[providerName]
-	state.registered = true
-	state.lastAvailabilityCheckAt = time.Now().UTC()
-	if err != nil {
-		state.lastAvailabilityError = failureMessage(err.Error())
-	} else {
-		state.lastAvailabilityOKAt = state.lastAvailabilityCheckAt
-		state.lastAvailabilityError = ""
-	}
-	r.providerRuntime[providerName] = state
-}
-
-// markProviderInventoryStale flags providerName's carried inventory as stale
-// after a failed live probe, but only while at least one other registered
-// provider is healthy. Without a healthy alternative, skipping the provider
-// would leave virtual-model resolution with no target at all (a 404 for the
-// alias) — routing to it and failing with the provider's 502/503 is the more
-// honest degradation, and it keeps single-provider deployments and
-// control-plane-only outages routable.
-func (r *ModelRegistry) markProviderInventoryStale(providerName string) {
-	providerName = strings.TrimSpace(providerName)
-	if providerName == "" {
-		return
-	}
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	if len(r.modelsByProvider[providerName]) == 0 || r.providerRuntime[providerName].inventoryStale {
-		return
-	}
-
-	healthyAlternative := false
-	for _, provider := range r.providers {
-		name := r.providerNames[provider]
-		if name == "" || name == providerName {
-			continue
-		}
-		state := r.providerRuntime[name]
-		if state.registered && !state.inventoryStale &&
-			state.lastModelFetchError == "" &&
-			state.lastAvailabilityError == "" &&
-			len(r.modelsByProvider[name]) > 0 {
-			healthyAlternative = true
-			break
-		}
-	}
-	if !healthyAlternative {
-		return
-	}
-
-	state := r.providerRuntime[providerName]
-	state.inventoryStale = true
-	r.providerRuntime[providerName] = state
-	r.models = rebuildGlobalModelMap(r.modelsByProvider, r.freshFirstProviderOrderLocked())
-	r.invalidateSortedCaches()
-}
-
-// FailedProviderNames returns configured provider names whose latest model
-// refresh attempt or availability probe failed. The background recheck loop
-// uses this to re-probe only the providers that are currently down. The
-// availability side matters: a request-time refresh can fail the availability
-// gate (marking the inventory stale) without ever attempting a model fetch,
-// and such a provider must still be re-probed to detect its recovery.
-func (r *ModelRegistry) FailedProviderNames() []string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	names := make([]string, 0)
-	for _, provider := range r.providers {
-		providerName := r.providerNames[provider]
-		if providerName == "" {
-			continue
-		}
-		state := r.providerRuntime[providerName]
-		if state.lastModelFetchError == "" && state.lastAvailabilityError == "" {
-			continue
-		}
-		names = append(names, providerName)
-	}
-	return names
-}
-
-// ProviderRuntimeSnapshots returns runtime diagnostics for configured providers
-// keyed by configured provider name.
-func (r *ModelRegistry) ProviderRuntimeSnapshots() []ProviderRuntimeSnapshot {
-	r.mu.RLock()
-	result := make([]ProviderRuntimeSnapshot, 0, len(r.providers))
-	for _, provider := range r.providers {
-		providerName := r.providerNames[provider]
-		if providerName == "" {
-			continue
-		}
-		state := r.providerRuntime[providerName]
-		result = append(result, ProviderRuntimeSnapshot{
-			Name:                    providerName,
-			Type:                    r.providerTypes[provider],
-			Registered:              state.registered,
-			DiscoveredModelCount:    len(r.modelsByProvider[providerName]),
-			LastModelFetchAt:        timePtrUTC(state.lastModelFetchAt),
-			LastModelFetchSuccessAt: timePtrUTC(state.lastModelFetchSuccessAt),
-			LastModelFetchError:     state.lastModelFetchError,
-			LastAvailabilityCheckAt: timePtrUTC(state.lastAvailabilityCheckAt),
-			LastAvailabilityOKAt:    timePtrUTC(state.lastAvailabilityOKAt),
-			LastAvailabilityError:   state.lastAvailabilityError,
-			InventoryStale:          state.inventoryStale,
-		})
-	}
-	r.mu.RUnlock()
-
-	initialized := r.IsInitialized()
-	for i := range result {
-		result[i].RegistryInitialized = initialized
-		result[i].UsingCachedModels = result[i].DiscoveredModelCount > 0 &&
-			!initialized &&
-			result[i].LastModelFetchSuccessAt == nil
-	}
-
-	return result
 }
