+++ internal/server/handlers.go
@@ -3,12 +3,8 @@ package server
 
 import (
 	"net/http"
-	"strings"
 	"sync"
 
-	"github.com/labstack/echo/v5"
-
-	"github.com/enterpilot/gomodel/internal/anthropicapi"
 	"github.com/enterpilot/gomodel/internal/auditlog"
 	batchstore "github.com/enterpilot/gomodel/internal/batch"
 	"github.com/enterpilot/gomodel/internal/conversationstore"
@@ -311,817 +307,4 @@ func (h *Handler) passthrough() *passthroughService {
 		normalizePassthroughV1Prefix: h.normalizePassthroughV1Prefix,
 		enabledPassthroughProviders:  h.enabledPassthroughProviders,
 	}
-}
-
-// ProviderPassthrough handles opaque provider-native requests under /p/{provider}/{endpoint}.
-//
-// OpenAI and Anthropic are the first-class providers in this ADR-0002 slice. Other
-// providers are intentionally deferred until they fit the same low-friction opaque path.
-//
-// @Summary      Provider passthrough
-// @Description  Runtime-configurable passthrough endpoint under /p/{provider}/{endpoint}; enabled by default via server.enable_passthrough_routes. The endpoint path is opaque and may proxy JSON, binary, or SSE responses with upstream status codes preserved. For multi-segment provider endpoints, clients that rely on OpenAPI-generated path handling should URL-encode embedded slashes in the endpoint parameter. A leading v1/ segment is normalized away by default so /p/{provider}/v1/... and /p/{provider}/... map to the same upstream path relative to the provider base URL.
-// @Tags         passthrough
-// @Accept       json
-// @Accept       mpfd
-// @Produce      json
-// @Produce      application/octet-stream
-// @Produce      text/event-stream
-// @Security     BearerAuth
-// @Param        provider  path      string  true  "Provider type"
-// @Param        endpoint  path      string  true  "Provider-native endpoint path relative to the provider base URL. URL-encode embedded / characters when using generated clients."
-// @Success      200       {file}    file    "Opaque upstream response body"
-// @Success      201       {file}    file    "Opaque upstream response body"
-// @Success      202       {file}    file    "Opaque upstream response body"
-// @Success      204       {string}  string  "No Content passthrough response"
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /p/{provider}/{endpoint} [get]
-// @Router       /p/{provider}/{endpoint} [post]
-// @Router       /p/{provider}/{endpoint} [put]
-// @Router       /p/{provider}/{endpoint} [patch]
-// @Router       /p/{provider}/{endpoint} [delete]
-// @Router       /p/{provider}/{endpoint} [head]
-// @Router       /p/{provider}/{endpoint} [options]
-func (h *Handler) ProviderPassthrough(c *echo.Context) error {
-	// A websocket upgrade on a passthrough route is a realtime session, not an
-	// HTTP proxy request; relay it through the realtime service instead.
-	if isWebSocketUpgrade(c.Request()) {
-		providerType, _, endpoint, _, err := passthroughExecutionTarget(c, h.provider, h.normalizePassthroughV1Prefix)
-		if err != nil {
-			return handleError(c, err)
-		}
-		// Realtime upgrades honor the same provider allowlist as the HTTP
-		// passthrough path: a provider disabled for /p/{provider}/... must not be
-		// reachable via a websocket upgrade.
-		if !isEnabledPassthroughProvider(providerType, h.enabledPassthroughProviders) {
-			return handleError(c, h.passthrough().unsupportedPassthroughProviderError(providerType))
-		}
-		// endpoint may carry the query string (e.g. "realtime?model=..."); compare
-		// only the path segment. The translations path selects the translation
-		// session surface, exactly as the typed route does.
-		endpointPath := strings.Trim(strings.SplitN(endpoint, "?", 2)[0], "/")
-		intent := ""
-		switch endpointPath {
-		case "realtime":
-		case "realtime/translations":
-			intent = core.RealtimeIntentTranslation
-		default:
-			return handleError(c, core.NewNotFoundError("unsupported realtime passthrough endpoint: "+endpointPath))
-		}
-		return h.realtime().PassthroughRealtime(c, providerType, intent)
-	}
-	return h.passthrough().ProviderPassthrough(c)
-}
-
-// Realtime handles GET /v1/realtime.
-//
-// @Summary      Open a realtime session
-// @Description  Upgrades to a websocket and relays an OpenAI-compatible realtime (speech-to-speech) session to the provider that owns the model named in the ?model= query parameter. Provider credentials are injected by the gateway. Passing ?call_id= instead attaches to an existing WebRTC/SIP call as a sideband channel; calls created through this gateway instance are routed automatically, others need explicit model (and provider) parameters.
-// @Tags         realtime
-// @Security     BearerAuth
-// @Param        model     query     string  false  "Model that owns the realtime session (required unless call_id names a call created through this gateway instance)"
-// @Param        provider  query     string  false  "Optional provider hint"
-// @Param        call_id   query     string  false  "Existing WebRTC/SIP call to attach to as a sideband channel"
-// @Success      101       {string}  string  "Switching Protocols"
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      429       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/realtime [get]
-func (h *Handler) Realtime(c *echo.Context) error {
-	return h.realtime().Realtime(c)
-}
-
-// RealtimeTranslations handles GET /v1/realtime/translations.
-//
-// @Summary      Open a realtime translation session
-// @Description  Upgrades to a websocket and relays an OpenAI-compatible realtime speech translation session (e.g. gpt-realtime-translate) to the provider that owns the model named in the ?model= query parameter. Translation sessions use their own event namespace and set the target language with session.update (session.audio.output.language); provider credentials are injected by the gateway. Equivalent to /v1/realtime?intent=translation.
-// @Tags         realtime
-// @Security     BearerAuth
-// @Param        model     query     string  true   "Translation model that owns the session"
-// @Param        provider  query     string  false  "Optional provider hint"
-// @Success      101       {string}  string  "Switching Protocols"
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      429       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/realtime/translations [get]
-func (h *Handler) RealtimeTranslations(c *echo.Context) error {
-	return h.realtime().RealtimeTranslations(c)
-}
-
-// MCP handles the aggregated MCP endpoint at /mcp.
-//
-// @Summary      MCP gateway (aggregated)
-// @Description  Streamable-HTTP MCP endpoint aggregating every configured upstream MCP server visible to the caller. Tools and prompts are namespaced as {slug}_{name}. POST carries JSON-RPC messages, GET opens the server-notification SSE stream, DELETE ends the session. The X-MCP-Servers request header optionally narrows the visible servers to a comma-separated subset of server slugs.
-// @Tags         mcp
-// @Accept       json
-// @Produce      json
-// @Produce      text/event-stream
-// @Security     BearerAuth
-// @Success      200  {object}  map[string]interface{}  "JSON-RPC response or SSE stream"
-// @Failure      401  {object}  core.OpenAIErrorEnvelope
-// @Failure      429  {object}  core.OpenAIErrorEnvelope
-// @Failure      501  {object}  core.OpenAIErrorEnvelope
-// @Router       /mcp [post]
-// @Router       /mcp [get]
-// @Router       /mcp [delete]
-func (h *Handler) MCP(c *echo.Context) error {
-	return h.mcp().handle(c, "")
-}
-
-// MCPServer handles the per-server MCP endpoints at /mcp/{server}.
-//
-// @Summary      MCP gateway (single server)
-// @Description  Streamable-HTTP MCP endpoint exposing one configured upstream MCP server with original (un-prefixed) tool names.
-// @Tags         mcp
-// @Accept       json
-// @Produce      json
-// @Produce      text/event-stream
-// @Security     BearerAuth
-// @Param        server  path      string  true  "Configured MCP server slug"
-// @Success      200     {object}  map[string]interface{}  "JSON-RPC response or SSE stream"
-// @Failure      401     {object}  core.OpenAIErrorEnvelope
-// @Failure      404     {object}  core.OpenAIErrorEnvelope
-// @Failure      429     {object}  core.OpenAIErrorEnvelope
-// @Failure      501     {object}  core.OpenAIErrorEnvelope
-// @Router       /mcp/{server} [post]
-// @Router       /mcp/{server} [get]
-// @Router       /mcp/{server} [delete]
-func (h *Handler) MCPServer(c *echo.Context) error {
-	return h.mcp().handle(c, c.Param("server"))
-}
-
-// RealtimeCalls handles POST /v1/realtime/calls.
-//
-// @Summary      Create a realtime WebRTC call
-// @Description  OpenAI-compatible WebRTC SDP exchange. Accepts a raw application/sdp offer with the model in the ?model= query parameter, or a multipart form with sdp and session (JSON) fields. The gateway routes by model, injects provider credentials, and relays the SDP answer; the Location header carries the created call id. Media flows directly between the client and the provider, so usage is recorded by a gateway-side sideband observer when usage tracking is enabled.
-// @Tags         realtime
-// @Accept       application/sdp
-// @Accept       mpfd
-// @Produce      application/sdp
-// @Produce      json
-// @Security     BearerAuth
-// @Param        model     query     string  false  "Model that owns the call (required for application/sdp offers)"
-// @Param        provider  query     string  false  "Optional provider hint"
-// @Param        request   body      string  true   "SDP offer (raw application/sdp body), or a multipart form with sdp and session (JSON) fields"
-// @Success      201       {string}  string  "SDP answer"
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      429       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/realtime/calls [post]
-func (h *Handler) RealtimeCalls(c *echo.Context) error {
-	return h.realtime().RealtimeCalls(c)
-}
-
-// RealtimeTranslationCalls handles POST /v1/realtime/translations/calls.
-//
-// @Summary      Create a realtime translation call (WebRTC)
-// @Description  OpenAI-compatible WebRTC SDP exchange for speech translation sessions. Identical to /v1/realtime/calls except that the session runs on the provider's translation surface; the created call is addressed at /v1/realtime/translations/calls/{call_id}.
-// @Tags         realtime
-// @Accept       application/sdp
-// @Accept       mpfd
-// @Produce      application/sdp
-// @Produce      json
-// @Security     BearerAuth
-// @Param        model     query     string  false  "Translation model that owns the session (required for raw SDP offers)"
-// @Param        provider  query     string  false  "Optional provider hint"
-// @Param        request   body      string  true   "SDP offer (raw application/sdp body), or a multipart form with sdp and session (JSON) fields"
-// @Success      201       {string}  string  "SDP answer"
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      429       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/realtime/translations/calls [post]
-func (h *Handler) RealtimeTranslationCalls(c *echo.Context) error {
-	return h.realtime().RealtimeTranslationCalls(c)
-}
-
-// RealtimeClientSecrets handles POST /v1/realtime/client_secrets.
-//
-// @Summary      Mint an ephemeral realtime client secret
-// @Description  OpenAI-compatible ephemeral credential minting for browser and mobile realtime clients. Routes by session.model (or the transcription model for transcription sessions), applies the same model-access, budget, and rate-limit gates as other model endpoints, and relays the provider response verbatim. The minted secret authenticates the client directly against the provider, bypassing the gateway for the session itself.
-// @Tags         realtime
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        provider  query     string  false  "Optional provider hint"
-// @Param        request   body      object{session=object,expires_after=object}  true  "Client secret request: session config with the routing model (session.model, or the nested transcription model) plus optional expires_after; additional fields are relayed verbatim"
-// @Success      200       {object}  map[string]any  "Provider client secret response"
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      429       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/realtime/client_secrets [post]
-func (h *Handler) RealtimeClientSecrets(c *echo.Context) error {
-	return h.realtime().RealtimeClientSecrets(c)
-}
-
-// RealtimeTranslationClientSecrets handles POST /v1/realtime/translations/client_secrets.
-//
-// @Summary      Mint an ephemeral realtime translation client secret
-// @Description  OpenAI-compatible ephemeral credential minting for browser and mobile clients that open a speech translation session. Identical to /v1/realtime/client_secrets except that the credential is minted on the provider's translation surface.
-// @Tags         realtime
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        provider  query     string  false  "Optional provider hint"
-// @Param        request   body      object  true   "Realtime translation session config carrying session.model"
-// @Success      200       {object}  map[string]interface{}
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      429       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/realtime/translations/client_secrets [post]
-func (h *Handler) RealtimeTranslationClientSecrets(c *echo.Context) error {
-	return h.realtime().RealtimeTranslationClientSecrets(c)
-}
-
-// ChatCompletion handles POST /v1/chat/completions
-//
-// @Summary      Create a chat completion
-// @Tags         chat
-// @Accept       json
-// @Produce      json
-// @Produce      text/event-stream
-// @Security     BearerAuth
-// @Param        request  body      core.ChatRequest  true  "Chat completion request"
-// @Success      200      {object}  core.ChatResponse  "JSON response or SSE stream when stream=true"
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      429      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/chat/completions [post]
-func (h *Handler) ChatCompletion(c *echo.Context) error {
-	return h.translatedInference().ChatCompletion(c)
-}
-
-// Health handles GET /health
-//
-// @Summary      Health check
-// @Tags         system
-// @Produce      json
-// @Success      200  {object}  map[string]string
-// @Router       /health [get]
-func (h *Handler) Health(c *echo.Context) error {
-	return c.JSON(http.StatusOK, map[string]string{"status": "ok"})
-}
-
-// ListModels handles GET /v1/models
-//
-// @Summary      List available models
-// @Tags         models
-// @Produce      json
-// @Security     BearerAuth
-// @Success      200  {object}  core.ModelsResponse
-// @Failure      401  {object}  core.OpenAIErrorEnvelope
-// @Failure      502  {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/models [get]
-func (h *Handler) ListModels(c *echo.Context) error {
-	// /v1/models is not a model-interaction endpoint, so the snapshot
-	// middleware does not lift the user-path header into the context. A
-	// managed key's bound path (set by auth middleware) still wins; the header
-	// only fills in for header-only and master-key callers, so the list is
-	// filtered by the same user-path policies inference enforces.
-	if ok, err := applyUserPathHeaderToContext(c); !ok {
-		return err
-	}
-
-	// Create context with request ID for provider
-	requestID := c.Request().Header.Get(core.RequestIDHeader)
-	ctx := core.WithRequestID(c.Request().Context(), requestID)
-
-	resp, err := h.provider.ListModels(ctx)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if h.keepOnlyAliasesAtModelsEndpoint {
-		object := "list"
-		if resp != nil && resp.Object != "" {
-			object = resp.Object
-		}
-		resp = &core.ModelsResponse{Object: object, Data: []core.Model{}}
-	}
-	if h.modelAuthorizer != nil && resp != nil {
-		resp = &core.ModelsResponse{
-			Object: resp.Object,
-			Data:   h.modelAuthorizer.FilterPublicModels(c.Request().Context(), resp.Data),
-		}
-	}
-	if h.exposedModelLister != nil {
-		ctx := c.Request().Context()
-		// The target-access filter is only available when an authorizer is set.
-		var allow func(core.ModelSelector) bool
-		if h.modelAuthorizer != nil {
-			allow = func(selector core.ModelSelector) bool {
-				return h.modelAuthorizer.AllowsModel(ctx, selector)
-			}
-		}
-		// User-path scoping of redirects is a property of the redirect itself, not
-		// of the authorizer, so it must apply even when no authorizer is configured
-		// (allow is nil there) — otherwise scoped redirect IDs leak to callers
-		// outside their user_paths.
-		if scoped, ok := h.exposedModelLister.(UserPathExposedModelLister); ok {
-			resp = mergeExposedModelsResponse(resp, scoped.ExposedModelsForUserPath(core.UserPathFromContext(ctx), allow))
-		} else if filtered, ok := h.exposedModelLister.(FilteredExposedModelLister); ok && allow != nil {
-			resp = mergeExposedModelsResponse(resp, filtered.ExposedModelsFiltered(allow))
-		} else {
-			exposed := h.exposedModelLister.ExposedModels()
-			if allow != nil {
-				filtered := make([]core.Model, 0, len(exposed))
-				for _, model := range exposed {
-					selector, err := core.ParseModelSelector(model.ID, "")
-					if err != nil || !allow(selector) {
-						continue
-					}
-					filtered = append(filtered, model)
-				}
-				exposed = filtered
-			}
-			resp = mergeExposedModelsResponse(resp, exposed)
-		}
-	}
-
-	// The models route is shared by both wire dialects. Anthropic SDK clients
-	// are identified by the anthropic-version header they always send; render
-	// the Anthropic list shape for them, the OpenAI shape for everyone else.
-	if c.Request().Header.Get("anthropic-version") != "" {
-		var models []core.Model
-		if resp != nil {
-			models = resp.Data
-		}
-		return c.JSON(http.StatusOK, anthropicapi.FromModels(models))
-	}
-
-	return c.JSON(http.StatusOK, resp)
-}
-
-// CreateFile handles POST /v1/files.
-//
-// @Summary      Upload a file
-// @Tags         files
-// @Accept       mpfd
-// @Produce      json
-// @Security     BearerAuth
-// @Param        provider  query     string  false  "Provider override when multiple providers are configured"
-// @Param        purpose   formData  string  true   "File purpose"
-// @Param        file      formData  file    true   "File to upload"
-// @Success      200       {object}  core.FileObject
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/files [post]
-func (h *Handler) CreateFile(c *echo.Context) error {
-	return h.nativeFiles().CreateFile(c)
-}
-
-// ListFiles handles GET /v1/files.
-//
-// @Summary      List files
-// @Tags         files
-// @Produce      json
-// @Security     BearerAuth
-// @Param        provider  query     string  false  "Provider filter"
-// @Param        purpose   query     string  false  "File purpose filter"
-// @Param        after     query     string  false  "Pagination cursor"
-// @Param        limit     query     int     false  "Maximum items to return (1-100, default 20)"
-// @Success      200       {object}  core.FileListResponse
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/files [get]
-func (h *Handler) ListFiles(c *echo.Context) error {
-	return h.nativeFiles().ListFiles(c)
-}
-
-// GetFile handles GET /v1/files/{id}.
-//
-// @Summary      Get file metadata
-// @Tags         files
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id        path      string  true   "File ID"
-// @Param        provider  query     string  false  "Provider override"
-// @Success      200       {object}  core.FileObject
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/files/{id} [get]
-func (h *Handler) GetFile(c *echo.Context) error {
-	return h.nativeFiles().GetFile(c)
-}
-
-// DeleteFile handles DELETE /v1/files/{id}.
-//
-// @Summary      Delete a file
-// @Tags         files
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id        path      string  true   "File ID"
-// @Param        provider  query     string  false  "Provider override"
-// @Success      200       {object}  core.FileDeleteResponse
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/files/{id} [delete]
-func (h *Handler) DeleteFile(c *echo.Context) error {
-	return h.nativeFiles().DeleteFile(c)
-}
-
-// GetFileContent handles GET /v1/files/{id}/content.
-//
-// @Summary      Download file content
-// @Tags         files
-// @Produce      application/octet-stream
-// @Security     BearerAuth
-// @Param        id        path   string  true   "File ID"
-// @Param        provider  query  string  false  "Provider override"
-// @Success      200       {file}  file  "Raw file content"
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/files/{id}/content [get]
-func (h *Handler) GetFileContent(c *echo.Context) error {
-	return h.nativeFiles().GetFileContent(c)
-}
-
-// AudioSpeech handles POST /v1/audio/speech.
-//
-// @Summary      Create speech (text-to-speech)
-// @Tags         audio
-// @Accept       json
-// @Produce      application/octet-stream
-// @Security     BearerAuth
-// @Param        request  body      core.AudioSpeechRequest  true  "Text-to-speech request"
-// @Success      200      {file}    file  "Binary audio in the requested response_format"
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      404      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/audio/speech [post]
-func (h *Handler) AudioSpeech(c *echo.Context) error {
-	return h.audio().CreateSpeech(c)
-}
-
-// AudioTranscriptions handles POST /v1/audio/transcriptions.
-//
-// @Summary      Create transcription (speech-to-text)
-// @Tags         audio
-// @Accept       mpfd
-// @Produce      json
-// @Produce      plain
-// @Security     BearerAuth
-// @Param        file             formData  file    true   "Audio file to transcribe"
-// @Param        model            formData  string  true   "Model ID"
-// @Param        language         formData  string  false  "Input language (ISO-639-1)"
-// @Param        prompt           formData  string  false  "Optional text to guide the model"
-// @Param        response_format          formData  string    false  "json, text, srt, verbose_json, or vtt"
-// @Param        temperature              formData  number    false  "Sampling temperature (0-1)"
-// @Param        timestamp_granularities[] formData  []string  false  "Timestamp granularities to populate: word and/or segment"
-// @Success      200                      {object}  map[string]interface{}  "Transcription in the requested response_format: a JSON object for json/verbose_json, or a text/plain body for text/srt/vtt"
-// @Failure      400              {object}  core.OpenAIErrorEnvelope
-// @Failure      401              {object}  core.OpenAIErrorEnvelope
-// @Failure      404              {object}  core.OpenAIErrorEnvelope
-// @Failure      502              {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/audio/transcriptions [post]
-func (h *Handler) AudioTranscriptions(c *echo.Context) error {
-	return h.audio().CreateTranscription(c)
-}
-
-// ImageGenerations handles POST /v1/images/generations.
-//
-// @Summary      Create image (image generation)
-// @Tags         images
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        request  body      core.ImageGenerationRequest  true  "Image generation request"
-// @Success      200      {object}  core.ImageGenerationResponse
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      404      {object}  core.OpenAIErrorEnvelope
-// @Failure      429      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/images/generations [post]
-func (h *Handler) ImageGenerations(c *echo.Context) error {
-	return h.images().CreateImage(c)
-}
-
-// ImageEdits handles POST /v1/images/edits.
-//
-// @Summary      Create image edit (inpainting / image-to-image)
-// @Tags         images
-// @Accept       mpfd
-// @Produce      json
-// @Security     BearerAuth
-// @Param        image            formData  file    false  "Source image to edit (single-image form; gpt-image-1 and DALL·E 2). At least one of image or image[] is required"
-// @Param        image[]          formData  file    false  "Repeatable field carrying several source images (gpt-image-1, up to 16). At least one of image or image[] is required"
-// @Param        prompt           formData  string  true   "Text description of the desired edit"
-// @Param        model            formData  string  true   "Model ID"
-// @Param        mask             formData  file    false  "PNG whose transparent areas mark where the image should be edited"
-// @Param        n                formData  integer false  "Number of images to generate"  minimum(1)
-// @Param        size             formData  string  false  "Output size, e.g. 1024x1024"
-// @Param        quality          formData  string  false  "Output quality (model-specific)"
-// @Param        response_format  formData  string  false  "url or b64_json (DALL·E 2 only; gpt-image-1 always returns b64_json)"
-// @Param        user             formData  string  false  "End-user identifier forwarded to the provider"
-// @Success      200      {object}  core.ImageGenerationResponse
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      404      {object}  core.OpenAIErrorEnvelope
-// @Failure      429      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/images/edits [post]
-func (h *Handler) ImageEdits(c *echo.Context) error {
-	return h.images().CreateImageEdit(c)
-}
-
-// AudioTranslations handles POST /v1/audio/translations.
-//
-// @Summary      Translate audio into English
-// @Tags         audio
-// @Accept       mpfd
-// @Produce      json
-// @Produce      plain
-// @Security     BearerAuth
-// @Param        file             formData  file    true   "Audio file to translate"
-// @Param        model            formData  string  true   "Model ID"
-// @Param        prompt           formData  string  false  "Optional English text to guide the model"
-// @Param        response_format  formData  string  false  "json, text, srt, verbose_json, or vtt"
-// @Param        temperature      formData  number  false  "Sampling temperature (0-1)"
-// @Success      200              {object}  map[string]interface{}  "English translation in the requested response_format"
-// @Failure      400              {object}  core.OpenAIErrorEnvelope
-// @Failure      401              {object}  core.OpenAIErrorEnvelope
-// @Failure      404              {object}  core.OpenAIErrorEnvelope
-// @Failure      502              {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/audio/translations [post]
-func (h *Handler) AudioTranslations(c *echo.Context) error {
-	return h.audio().CreateTranslation(c)
-}
-
-// Responses handles POST /v1/responses
-//
-// @Summary      Create a model response (Responses API)
-// @Tags         responses
-// @Accept       json
-// @Produce      json
-// @Produce      text/event-stream
-// @Security     BearerAuth
-// @Param        request  body      core.ResponsesRequest  true  "Responses API request"
-// @Success      200      {object}  core.ResponsesResponse  "JSON response or SSE stream when stream=true"
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      429      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/responses [post]
-func (h *Handler) Responses(c *echo.Context) error {
-	return h.translatedInference().Responses(c)
-}
-
-// GetResponse handles GET /v1/responses/{id}.
-//
-// @Summary      Get a response
-// @Tags         responses
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id        path      string  true   "Response ID"
-// @Param        provider  query     string  false  "Provider override for native lookups"
-// @Param        include   query     []string false "Fields to include in the response" collectionFormat(multi)
-// @Param        include[] query     []string false "Fields to include in the response" collectionFormat(multi)
-// @Param        include_obfuscation query bool false "Whether to include obfuscated response data"
-// @Param        starting_after query int false "Input item offset for providers that support it"
-// @Success      200       {object}  core.ResponsesResponse
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/responses/{id} [get]
-func (h *Handler) GetResponse(c *echo.Context) error {
-	return h.nativeResponses().GetResponse(c)
-}
-
-// ListResponseInputItems handles GET /v1/responses/{id}/input_items.
-//
-// @Summary      List response input items
-// @Tags         responses
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id        path      string  true   "Response ID"
-// @Param        provider  query     string  false  "Provider override for native lookups"
-// @Param        after     query     string  false  "Pagination cursor"
-// @Param        include   query     []string false "Fields to include in listed input items" collectionFormat(multi)
-// @Param        include[] query     []string false "Fields to include in listed input items" collectionFormat(multi)
-// @Param        limit     query     int     false  "Maximum items to return (1-100, default 20)"
-// @Param        order     query     string  false  "Sort order: asc or desc"  Enums(asc, desc)
-// @Success      200       {object}  core.ResponseInputItemListResponse
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/responses/{id}/input_items [get]
-func (h *Handler) ListResponseInputItems(c *echo.Context) error {
-	return h.nativeResponses().ListResponseInputItems(c)
-}
-
-// CancelResponse handles POST /v1/responses/{id}/cancel.
-//
-// @Summary      Cancel a response
-// @Tags         responses
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id        path      string  true   "Response ID"
-// @Param        provider  query     string  false  "Provider override for native cancellation"
-// @Success      200       {object}  core.ResponsesResponse
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/responses/{id}/cancel [post]
-func (h *Handler) CancelResponse(c *echo.Context) error {
-	return h.nativeResponses().CancelResponse(c)
-}
-
-// DeleteResponse handles DELETE /v1/responses/{id}.
-//
-// @Summary      Delete a response
-// @Tags         responses
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id        path      string  true   "Response ID"
-// @Param        provider  query     string  false  "Provider override for native deletion"
-// @Success      200       {object}  core.ResponseDeleteResponse
-// @Failure      400       {object}  core.OpenAIErrorEnvelope
-// @Failure      401       {object}  core.OpenAIErrorEnvelope
-// @Failure      404       {object}  core.OpenAIErrorEnvelope
-// @Failure      501       {object}  core.OpenAIErrorEnvelope
-// @Failure      502       {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/responses/{id} [delete]
-func (h *Handler) DeleteResponse(c *echo.Context) error {
-	return h.nativeResponses().DeleteResponse(c)
-}
-
-// ResponseInputTokens handles POST /v1/responses/input_tokens.
-//
-// @Summary      Count response input tokens
-// @Tags         responses
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        request  body      core.ResponseInputTokensRequest  true  "Response input token request"
-// @Success      200      {object}  core.ResponseInputTokensResponse
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      501      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/responses/input_tokens [post]
-func (h *Handler) ResponseInputTokens(c *echo.Context) error {
-	return h.nativeResponses().CountResponseInputTokens(c)
-}
-
-// CompactResponse handles POST /v1/responses/compact.
-//
-// @Summary      Compact response input
-// @Tags         responses
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        request  body      core.ResponseCompactRequest  true  "Response compact request"
-// @Success      200      {object}  core.ResponseCompactResponse
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      501      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/responses/compact [post]
-func (h *Handler) CompactResponse(c *echo.Context) error {
-	return h.nativeResponses().CompactResponse(c)
-}
-
-// Embeddings handles POST /v1/embeddings
-//
-// @Summary      Create embeddings
-// @Tags         embeddings
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        request  body      core.EmbeddingRequest  true  "Embeddings request"
-// @Success      200      {object}  core.EmbeddingResponse
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      429      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/embeddings [post]
-func (h *Handler) Embeddings(c *echo.Context) error {
-	return h.translatedInference().Embeddings(c)
-}
-
-// Batches handles POST /v1/batches.
-//
-// OpenAI-compatible fields are accepted (`input_file_id`, `endpoint`, `completion_window`, `metadata`).
-// Inline `requests` are also accepted for providers with native inline batch support (for example Anthropic).
-//
-// @Summary      Create a native provider batch
-// @Tags         batch
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        request  body      core.BatchRequest  true  "Batch request"
-// @Success      200      {object}  core.BatchResponse
-// @Failure      400      {object}  core.OpenAIErrorEnvelope
-// @Failure      401      {object}  core.OpenAIErrorEnvelope
-// @Failure      502      {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/batches [post]
-func (h *Handler) Batches(c *echo.Context) error {
-	return h.nativeBatch().Batches(c)
-}
-
-// GetBatch handles GET /v1/batches/{id}.
-//
-// @Summary      Get a batch
-// @Tags         batch
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id   path      string  true  "Batch ID"
-// @Success      200  {object}  core.BatchResponse
-// @Failure      400  {object}  core.OpenAIErrorEnvelope
-// @Failure      401  {object}  core.OpenAIErrorEnvelope
-// @Failure      404  {object}  core.OpenAIErrorEnvelope
-// @Failure      500  {object}  core.OpenAIErrorEnvelope
-// @Failure      502  {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/batches/{id} [get]
-func (h *Handler) GetBatch(c *echo.Context) error {
-	return h.nativeBatch().GetBatch(c)
-}
-
-// ListBatches handles GET /v1/batches.
-//
-// @Summary      List batches
-// @Tags         batch
-// @Produce      json
-// @Security     BearerAuth
-// @Param        after  query     string  false  "Pagination cursor"
-// @Param        limit  query     int     false  "Maximum items to return (1-100, default 20)"
-// @Success      200    {object}  core.BatchListResponse
-// @Failure      400    {object}  core.OpenAIErrorEnvelope
-// @Failure      401    {object}  core.OpenAIErrorEnvelope
-// @Failure      404    {object}  core.OpenAIErrorEnvelope
-// @Failure      500    {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/batches [get]
-func (h *Handler) ListBatches(c *echo.Context) error {
-	return h.nativeBatch().ListBatches(c)
-}
-
-// CancelBatch handles POST /v1/batches/{id}/cancel.
-//
-// @Summary      Cancel a batch
-// @Tags         batch
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id   path      string  true  "Batch ID"
-// @Success      200  {object}  core.BatchResponse
-// @Failure      400  {object}  core.OpenAIErrorEnvelope
-// @Failure      401  {object}  core.OpenAIErrorEnvelope
-// @Failure      404  {object}  core.OpenAIErrorEnvelope
-// @Failure      500  {object}  core.OpenAIErrorEnvelope
-// @Failure      502  {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/batches/{id}/cancel [post]
-func (h *Handler) CancelBatch(c *echo.Context) error {
-	return h.nativeBatch().CancelBatch(c)
-}
-
-// BatchResults handles GET /v1/batches/{id}/results.
-//
-// @Summary      Get batch results
-// @Tags         batch
-// @Produce      json
-// @Security     BearerAuth
-// @Param        id   path      string  true  "Batch ID"
-// @Success      200  {object}  core.BatchResultsResponse
-// @Failure      400  {object}  core.OpenAIErrorEnvelope
-// @Failure      401  {object}  core.OpenAIErrorEnvelope
-// @Failure      404  {object}  core.OpenAIErrorEnvelope
-// @Failure      409  {object}  core.OpenAIErrorEnvelope
-// @Failure      500  {object}  core.OpenAIErrorEnvelope
-// @Failure      502  {object}  core.OpenAIErrorEnvelope
-// @Router       /v1/batches/{id}/results [get]
-func (h *Handler) BatchResults(c *echo.Context) error {
-	return h.nativeBatch().BatchResults(c)
 }
